==> 02_MergDischSampler_hydroAlteck2016.Rmd <==
---
title: "Merge Discharge, Rain & Sampler Data"
author: "PAZ"
date: "26 octobre 2016"
output: pdf_document
---

```{r, echo=FALSE, message=FALSE, include=FALSE}
Sys.setlocale("LC_ALL", "English")
```

## Purpose

A. This document merges corrected flowmeter data and automatic sampler data. 

B. Converts 6 min rainfall to monthly statistics with visualizations comparing discharge. 

Used files:

1. **hydroAlteck2016_smooth_R.csv**
2. **prelev_20160713.csv**

3. **sixMinutePluvioAlteck2016.csv**

4. **Data/2minRain.csv** (at the end)

Produced file:

1. **hydroAlteck2016_R.csv** (Used for plotting Sample and Discharge data together).
2. **RainAprJune2016_R.csv** (March and July delted)
3. **WaterDay_R.csv** (Daily and cumulative rainfall - mm)
4- **SamplerFormatted_R.csv** (Cleaned sampler)

## Required R-packages:

```{r, message=FALSE}

# Plotting functions
library("ggplot2")
library("scales")
library("tidyr")
library("dplyr")
library("reshape")
library("ggrepel")

```

## Working directory

```{r, message=FALSE}

# setwd("D:/Documents/these_pablo/Alteckendorf2016/R")
# setwd("/Users/DayTightChunks/Documents/PhD/Routput/Alteck/R")
# setwd("D:/Documents/these_pablo/Alteckendorf2016/00_TransparencyFolder/Discharge")
getwd()

```


## Import imputed discharge and rainfall data 

```{r, message=FALSE}

# Discharge
dischargeAlteck = read.csv2("Data/hydroAlteck2016_smooth_R.csv")
head(dischargeAlteck)

dischargeAlteck$Date = as.POSIXct(strptime(dischargeAlteck$DateCheck, 
                                           "%d/%m/%Y %H:%M"
                                           , tz="EST")
                                  )

dischargeAlteck$DayMoYr = as.POSIXct(strptime(dischargeAlteck$DateCheck, 
                                           "%d/%m/%Y"
                                           , tz="EST")
                                  )

sum(is.na(dischargeAlteck$Date))
naDates = dischargeAlteck[is.na(dischargeAlteck$Date == TRUE),]

duplicateAlteck <- dischargeAlteck[duplicated(dischargeAlteck$DateCheck),]
head(duplicateAlteck)

```
## Convert discharge $Q$ to volume $V$

Convert discharge at $\Delta t$ of 2 min:

$$ V(t) = \int^{\Delta t}_0 Q(t) dt $$
```{r}
## Convert m3.h -> m3
dischargeAlteck$Vol2min <- dischargeAlteck$Q.HW1*2/60
```


## Compare discharge to rainfall

```{r}
# Rainfall
rain = read.csv2("Data/sixMinutePluvioAlteck2016.csv", header = F)
head(rain)

rain$V1 <- as.character(rain$V1)
rain$Date = as.POSIXct(strptime(rain$V1,
                                "%d/%m/%Y %H:%M", tz="EST") )

rain$DayMoYr = as.POSIXct(strptime(rain$V1,
                                   "%d/%m/%Y", tz="EST")  ) 

sum(is.na(rain$Date))
naDates = rain[is.na(rain$Date == TRUE),]

#library(dplyr)
# detach("package:plyr")
dischDay <- dischargeAlteck %>%
  group_by(DayMoYr) %>%
  dplyr::summarize(Disch.mm = (sum(Vol2min)/(47*10000))*10^3 ) # Discharge in mm 

rainDay <- rain %>%
  group_by(DayMoYr) %>%
  dplyr::summarize(Rain.mm = sum(V2))

rainDay$CumRain <- cumsum(rainDay$Rain.mm)

waterDay <- merge(rainDay, dischDay, by = "DayMoYr")

waterDay$Month <- 
  ifelse(waterDay$DayMoYr >= as.POSIXct("2016-03-24 00:30:00", tz = "EST") &
           waterDay$DayMoYr < as.POSIXct("2016-04-01 00:00:00", tz = "EST"), "March",
         ifelse(waterDay$DayMoYr >= as.POSIXct("2016-04-01 00:00:00", tz = "EST") &
                  waterDay$DayMoYr < as.POSIXct("2016-05-01 00:00:00", tz = "EST"), "April",
                ifelse(waterDay$DayMoYr >= as.POSIXct("2016-05-01 00:00:00", tz = "EST") &
                         waterDay$DayMoYr < as.POSIXct("2016-06-01 00:00:00", tz = "EST"), "May",
                       ifelse(waterDay$DayMoYr >= as.POSIXct("2016-06-01 00:00:00", tz = "EST") & 
                                waterDay$DayMoYr < as.POSIXct("2016-07-01 00:00:00", tz = "EST"), "June", "July" )
                            )
                     )
         )
waterDay$CumRain.mm <- cumsum(waterDay$Rain.mm)

write.csv2(waterDay, 
           'Data/WaterDay_R.csv', row.names = F)

```

## Delete unused months

```{r}
waterDay <- subset(waterDay, Month != "March" & Month != "July")
waterDay$Month <- factor(waterDay$Month, levels = c("April" , "May", "June" ))
levels(waterDay$Month)

meltWaterDay <- melt(waterDay, id=c("DayMoYr", "Month"))

split <- strsplit(as.character(meltWaterDay$DayMoYr), "2016-", fixed = TRUE)
meltWaterDay$Day <- sapply(split, "[", 2)

meltWaterDay <- subset(meltWaterDay, variable != 'CumRain.mm')

# Test function
#g_param = 1.5
# g_param = 2.2  #  (Hoaglin et al.,1986; Hoaglin & Iglewicz, 1987) 
g_param = 4  #  Just to capture XXX events 
is_outlier <- function(x) {
  return(x < quantile(x, 0.25) - g_param * IQR(x) | x > quantile(x, 0.75) + g_param * IQR(x))
}

meltWaterDay2 <- meltWaterDay %>%
  group_by(Month) %>%
  #mutate(outlier = ifelse(is_outlier(value) & variable == "Disch.mm", as.character(Day), NA))
  mutate(outlier = ifelse(is_outlier(value), as.character(Day), NA))

hydroBox <- ggplot(meltWaterDay2) +
  geom_boxplot(data = subset(meltWaterDay2, variable != "CumRain"), aes(x = factor(Month), y = value, fill = variable)) + 
  theme_bw() +
  ylab("mm.day") +
  theme(axis.title.x = element_blank() ) +
  geom_text_repel(data = subset(meltWaterDay2, variable != "CumRain"), aes(x = factor(Month), y = value, label = outlier), 
                  size = 3,
                  arrow = arrow(length = unit(0.005, 'npc'), type = "closed"),
                  force = 0.5, 
                  point.padding = unit(0.9, 'lines'), 
                  max.iter = 2e3,
                  nudge_x = .05,
                  na.rm = TRUE) 
hydroBox
#ggsave(hydroBox, filename = "RainDischBox.png", width = 8, height = 5, units = "in", scale = 1)
```



## As Bar Plots

```{r}

dischDay$Month <- 
  ifelse(dischDay$DayMoYr >= as.POSIXct("2016-03-24 00:30:00", tz = "EST") &
           dischDay$DayMoYr < as.POSIXct("2016-04-01 00:00:00", tz = "EST"), "March",
         ifelse(dischDay$DayMoYr >= as.POSIXct("2016-04-01 00:00:00", tz = "EST") &
                  dischDay$DayMoYr < as.POSIXct("2016-05-01 00:00:00", tz = "EST"), "April",
                ifelse(dischDay$DayMoYr >= as.POSIXct("2016-05-01 00:00:00", tz = "EST") &
                         dischDay$DayMoYr < as.POSIXct("2016-06-01 00:00:00", tz = "EST"), "May",
                       ifelse(dischDay$DayMoYr >= as.POSIXct("2016-06-01 00:00:00", tz = "EST") &
                                dischDay$DayMoYr < as.POSIXct("2016-07-01 00:00:00", tz = "EST"), "June", "July" )
                       )
                )
         )

dischSumm <- dischDay %>%
  group_by(Month) %>%
  dplyr::summarize(Qtot.measure = sum(Disch.mm),
            Qmedian.measure = median(Disch.mm),
            Qmean.measure = mean(Disch.mm), 
            Qmean.SD = sd(Disch.mm))
  
rainDay$Month <- 
  ifelse(rainDay$DayMoYr >= as.POSIXct("2016-03-24 00:30:00", tz = "EST") &
           rainDay$DayMoYr < as.POSIXct("2016-04-01 00:00:00", tz = "EST"), "March",
         ifelse(rainDay$DayMoYr >= as.POSIXct("2016-04-01 00:00:00", tz = "EST") &
                  rainDay$DayMoYr < as.POSIXct("2016-05-01 00:00:00", tz = "EST"), "April",
                ifelse(rainDay$DayMoYr >= as.POSIXct("2016-05-01 00:00:00", tz = "EST") &
                         rainDay$DayMoYr < as.POSIXct("2016-06-01 00:00:00", tz = "EST"), "May",
                       ifelse(rainDay$DayMoYr >= as.POSIXct("2016-06-01 00:00:00", tz = "EST") &
                                rainDay$DayMoYr < as.POSIXct("2016-07-01 00:00:00", tz = "EST"), "June", "July" )
                       )
                )
         )

rainSumm <- rainDay %>%
  group_by(Month) %>%
  dplyr:: summarize(Rtot.measure = sum(Rain.mm), 
            Rmedian.measure = median(Rain.mm),
            Rmean.measure = mean(Rain.mm),
            Rmean.SD = sd(Rain.mm))

MonthSumm <- merge(rainSumm, dischSumm, by = "Month")
#MonthSumm$Rest.measure <- (MonthSumm$Qtot.measure/MonthSumm$Rtot.measure)*100

MonthSumm <- subset(MonthSumm, Month != "March" & Month != "July")


monthTidy <- MonthSumm %>%
  gather(measure, value, -Month) %>% # Melts data frame
  separate(measure, into = c("Source", "temporary_var")) %>% # parses the sep = "." into...
  spread(temporary_var, value) # Moves molten temporary variable to own column 

monthTidy$Type <- ifelse(monthTidy$Source == "Rtot" | monthTidy$Source == "Qtot", "Total", "Summary")
monthTidy <- subset(monthTidy, Source != "Qmedian" & Source != "Rmedian")

monthTidy$Month <- as.factor(monthTidy$Month)
levels(monthTidy$Month)
monthTidy$Month <- factor(monthTidy$Month, levels = c("April" , "May", "June" ))
monthTidy$Source <- factor(monthTidy$Source, levels = c("Rmean", "Qmean", "Rtot",  "Qtot"))
levels(monthTidy$Source)

ggplot(data = monthTidy,  aes(x=Month, y=measure, fill = Source )) +
  geom_bar(stat = "identity", position = "dodge", width = 0.5) + #, ymin=measure-SD, ymax=measure+SD) +
  facet_wrap(~Type, scale="free") +
  theme_bw() +
  geom_linerange(aes(ymin = measure, ymax = measure+SD),
                 width=.2 , # ) + #,                    # Width of the error bars
                  position=position_dodge(.5)) +
  ylab("mm") +
  theme(axis.title.x = element_blank() )
  # xlab("Month") #+
  #scale_fill_manual(#values = c("#6a51a3" , "#ec7014", "#807dba", "#fe9929"), # purple-orange
  #                  # values = c("#6a51a3" , "#ec7014", "#d9d9d9", "#fe9929"), # Unknown as grey
  #                  name= element_blank(), #"Mass Balance", # \n
  #                 breaks=c("Qmean", "Qtot" , "Rmean" , "Rtot"),
  #                  labels=c("Qmean", "Qtot" , "Rmean" , "Rtot")) +
  # guides(fill=guide_legend(ncol=2))
```



## Import raw sampler data (March 25th to Jul 12th)

```{r, message=FALSE}

samplesAlteck = read.csv2("Data/prelev_20160713.csv", header = FALSE)

head(samplesAlteck)


samplesAlteck = samplesAlteck[samplesAlteck$V2 != 0, ]
samplesAlteck$Date = as.POSIXct(strptime(samplesAlteck$V1, 
                                         "%d/%m/%Y %H:%M", 
                                         tz="EST"))
sum(is.na(samplesAlteck$V1))

samplesAlteck = samplesAlteck[,c(3,1:2)]
colnames(samplesAlteck) <- c("Date", "DateCheck", "sampleQ")

sum(is.na(samplesAlteck$Date))

samplesAlteck = samplesAlteck[order(samplesAlteck$Date),]

head(samplesAlteck)


write.csv2(samplesAlteck, 
           'Data/SamplerFormatted_R.csv', row.names = F)


```


## Merge the Discharge and the Samples' dataframes

To merge the two data.frames, we need to correct minutes in the sample data, some of which took place during odd minutes. 

1. Identify the odd minutes in a temporary data set to discard

```{r, message=FALSE}

discard = merge(dischargeAlteck, samplesAlteck, by = "Date", all = T)

# How many missing Discharge values resulting from the merge?
sum(is.na(discard$Date))
sum(is.na(discard$Q.m3Hrs))
naQs = discard[is.na(discard$Q.m3Hrs == TRUE),]

naQs$Date = naQs$Date+60

naQs = naQs[,c("Date", "DateCheck.y")]

head(naQs)
head(dischargeAlteck)




```

2. Add these odd-date markers to the flow-meter data (note that Date column remains as even minutes)

```{r, message=FALSE}

# Merge new dates to discharge data 
hydroAlteck2016 = merge(dischargeAlteck, naQs, by = c("Date"), all = T)
head(hydroAlteck2016)

# Check number of odd-minute dates, should be 0:
sum(is.na(hydroAlteck2016$Q.m3Hrs))

# Fill in the rest  of the Target dates (even)
hydroAlteck2016$DateCheck.S <- ifelse(is.na(hydroAlteck2016$DateCheck.y), 
                                      as.character(hydroAlteck2016$DateCheck), 
                                      as.character(hydroAlteck2016$DateCheck.y))
hydroAlteck2016$DateCheck.y <- NULL

# Create common column name in samples' target column (i.e. DateCheck.S)
samplesAlteck <- samplesAlteck[, c("DateCheck", "sampleQ")]
colnames(samplesAlteck) <- c("DateCheck.S", "sampleQ")
head(samplesAlteck)

```



3. Merging the two tables

```{r, message=FALSE}

hydroAlteck2016 = merge(hydroAlteck2016, samplesAlteck, by = c("DateCheck.S"), all = T)

# Checks
sum(is.na(hydroAlteck2016$Date))
anyDuplicated(hydroAlteck2016$Date)

sum(is.na(hydroAlteck2016$Q.m3Hrs))
head(hydroAlteck2016)

class(hydroAlteck2016$Date)

# Order by date
hydroAlteck2016 = hydroAlteck2016[order(hydroAlteck2016$Date),]

```

### Create a "Type"" column to point to Sampling times during plotting

```{r, message=FALSE}

hydroAlteck2016$Type = ifelse(is.na(hydroAlteck2016$sampleQ), "Discharge", "Sample")

head(hydroAlteck2016)
sum(is.na(hydroAlteck2016$Q.HW1))




```

## Plot a discharge graph with samples
```{r}
# hy =  subset(hydroAlteck2016, Date >=  & Date <= )
```


### Merge rainfall data

1. Merge 2 and 12 min minute pluvio data, handling NA's as 0 
```{r}

rain2min =  read.csv2("Data/2minRain.csv", header = T, dec = ".")

rain2min$Rain.mm <- as.numeric(rain2min$Rain.mm)
rain2min$Date <- as.character(rain2min$Date)
rain2min$Date <- as.POSIXct(strptime(rain2min$Date, 
                                         "%d/%m/%Y %H:%M", 
                                         tz="EST"))

hydroAlteck2016 = merge(hydroAlteck2016, rain2min, by = "Date", all = T)
hydroAlteck2016$Rain.mm <- ifelse(is.na(hydroAlteck2016$Rain.mm), 0, hydroAlteck2016$Rain.mm)

rain12m = read.csv2("Data/12minRain.csv", dec = ".")
rain12m$Date = as.POSIXct(strptime(rain12m$Date,
                                "%d/%m/%Y %H:%M", tz="EST") )

colnames(rain12m)[2]  <- "Rain12min.mm"
hydroAlteck2016 <-  merge(hydroAlteck2016, rain12m, by = "Date", all = T)
hydroAlteck2016$Rain12min.mm <- 
  ifelse(is.na(hydroAlteck2016$Rain12min.mm), 0, hydroAlteck2016$Rain12min.mm)

# Should be zero
sum(is.na(hydroAlteck2016$Rain.mm))

# Check which rainfall minutes do not have Discharge, and if Rain = 0, delete obs.
# naValues <- subset(hydroAlteck2016, is.na(hydroAlteck2016$Q.HW1)) # Only NA's where Rain == 0 
hydroAlteck2016 <- subset(hydroAlteck2016, !is.na(hydroAlteck2016$Q.HW1))
```


## Saving

```{r}
write.csv2(hydroAlteck2016, "Data/hydroAlteck2016_R.csv", row.names = F)
```
