==> 02.1_Hyetograph.Rmd <==
---
title: "Hyetograph"
author: "PAZ"
date: "6 septembre 2017"
output: pdf_document
---

```{r, echo=FALSE, message=FALSE, include=FALSE}
Sys.setlocale("LC_ALL", "English")
```

## Purpose

This produces a Hyetograph (rainfall and discharge data) based on 2 min intrevals. 

Input files:

- **groupAlteck2016_R** (Contains discharge and rainfall @ 2 min)

Output files:

- **nothing for now** 

## Libraries
```{r}
# Hyetograph
library(ggplot2)

# getwd()
# setwd("D:/Documents/these_pablo/Alteckendorf2016/HydrologicalMonitoring")
```

## Data

```{r}
# Discharge and 2 min rainfall 
hydro = read.csv2("Data/groupAlteck2016_R.csv")
hydro$Date = as.POSIXct(strptime(hydro$DateCheck.S,
                                 "%d/%m/%Y %H:%M", tz="EST") )

names(hydro)
qra <- hydro[, c("Date", "Q.HW1", "Rain.mm")]

# Catchment area
area <- 47*10**4 # [m2]
```

## Conversions

Convert rainfall data [$mm$] to the same units as discharge [$m^3/h$]

```{r}
qra$precip_m3 = qra$Rain.mm/10^3 * area

# Calculate the range needed to avoid having your hyetograph and hydrograph overlap 
maxRange <- 1*(max(qra$precip_m3) + max(qra$Q.HW1))

# Create a function to backtransform the axis labels for precipitation
precip_labels <- function(x) {round( ((x / area) * 10^3), digits = 2) } # X will be precip_m3 -> convert to mm

# Plot the data
ggplot(data = qra,
       aes(x = Date)) +
  theme_minimal() +
  # Use geom_tile to create the inverted hyetograph. geom_tile has a bug that displays a warning message for height and width, you can ignore it.
  geom_tile(aes(y = -1*(precip_m3/2-maxRange), # y = the center point of each bar, as geom_tile uses variable as center point
                height = precip_m3,
                width = 1
                ),
            fill = "gray50",
            color = "forestgreen") +
  # Plot your discharge data
  geom_line(aes(y = Q.HW1),
            color = "blue") +
  # Create a second axis with sec_axis() and format the labels to display the original precipitation units.
  scale_y_continuous(name = "Discharge (m3/h)",
                     sec.axis = sec_axis(trans = ~-1*(.-maxRange), # Equivalent to: y2 = -1*(y1 - maxRange) 
                                         name = "Precipitation (mm)",
                                         labels = precip_labels)) # x here is = precip_cuft
```

## Rainfall alone (12 min)

```{r}
rain12m = read.csv2("Data/12minRain.csv", dec = ".")
rain12m$Date = as.POSIXct(strptime(rain12m$Date,
                                "%d/%m/%Y %H:%M", tz="EST") )

ggplot(data = rain12m, aes(x = Date)) + 
  geom_line(aes(y = Rain.mm))
  
  #geom_bar(aes(y = Rain.mm), stat="identity") + 
  #  scale_x_date(date_breaks = "1 month", 
  #               labels = date_format("%y/%m"),
  #               limits = dtLimits)  +
  #  theme(axis.text.x = element_text(angle = 90, vjust = .5))

```

# Average intensity at 30 min rainfall 

```{r}
rain30 <- read.csv2("Data/30minRain.csv", dec = ".")
# head(rain30)

rain30$Date = as.POSIXct(strptime(rain30$Time,
                                 "%d/%m/%Y %H:%M", tz="EST") )

rain30$DayMoYr = as.POSIXct(strptime(rain30$Time,
                                   "%d/%m/%Y", tz="EST")  ) 


rain30$Month <- 
  ifelse(rain30$DayMoYr >= as.POSIXct("2016-03-24 00:30:00", tz = "EST") &
           rain30$DayMoYr < as.POSIXct("2016-04-01 00:00:00", tz = "EST"), "March",
         ifelse(rain30$DayMoYr >= as.POSIXct("2016-04-01 00:00:00", tz = "EST") &
                  rain30$DayMoYr < as.POSIXct("2016-05-01 00:00:00", tz = "EST"), "April",
                ifelse(rain30$DayMoYr >= as.POSIXct("2016-05-01 00:00:00", tz = "EST") &
                         rain30$DayMoYr < as.POSIXct("2016-06-01 00:00:00", tz = "EST"), "May",
                       ifelse(rain30$DayMoYr >= as.POSIXct("2016-06-01 00:00:00", tz = "EST") & 
                                rain30$DayMoYr < as.POSIXct("2016-07-01 00:00:00", tz = "EST"), "June", "July" )
                            )
                     )
         )

rain30$mm.hr = rain30$Cumm.mm*(60/30)

# Delete all days with 0 rainfall
r<-rain30[!(rain30$mm.hr==0),]

library("dplyr")
rain <- r %>%
  group_by(Month) %>%
  dplyr::summarize(Int.mm.hr = mean(mm.hr))

rainSD <- r %>%
  group_by(Month) %>%
  dplyr::summarize(Int.mm.hr = sd(mm.hr))

rainSum <- r %>%
  group_by(Month) %>%
  dplyr::summarize(CummRain = sum(Cumm.mm))
```

# Discharge summary

```{r}
head(qra)

qra$Month <- 
  ifelse(qra$Date >= as.POSIXct("2016-03-24 00:30:00", tz = "EST") &
           qra$Date < as.POSIXct("2016-04-01 00:00:00", tz = "EST"), "March",
         ifelse(qra$Date >= as.POSIXct("2016-04-01 00:00:00", tz = "EST") &
                  qra$Date < as.POSIXct("2016-05-01 00:00:00", tz = "EST"), "April",
                ifelse(qra$Date >= as.POSIXct("2016-05-01 00:00:00", tz = "EST") &
                         qra$Date < as.POSIXct("2016-06-01 00:00:00", tz = "EST"), "May",
                       ifelse(qra$Date >= as.POSIXct("2016-06-01 00:00:00", tz = "EST") & 
                                qra$Date < as.POSIXct("2016-07-01 00:00:00", tz = "EST"), "June", "July" )
                            )
                     )
         )

qra$VolQ = qra$Q.HW1/60*2 # * 2 min / 60min = m3
qra$VolR = qra$Rain.mm/10^3*area # * m * m2 

qr = qra %>% 
  group_by(Month) %>%
  select(VolQ, VolR) %>%
  summarise_each(funs(sum = sum)) 

qr$Qmonth = qr$VolQ_sum/30
qr$Restit = (qr$VolQ_sum/qr$VolR_sum)*100

qr2 = qra %>% 
  group_by(Month) %>%
  select(VolQ, VolR) %>%
  summarise_each(funs(mean = mean, 
                      sd = sd))



```
