==> 04_WeeklyFlux_fluxAlteck2016.Rmd <==
---
title: "Weekly Flux Characterisitics"
author: "PAZ"
date: "26 octobre 2016"
output: pdf_document
---


```{r, echo=FALSE, message=FALSE, include=FALSE}
Sys.setlocale("LC_ALL", "English")
```

## Purpose

This document creates summary variables for discharge characteristics by sub-weeks. 

Input files:

- **hydroAlteck2016_R.csv** (Book 2.0)
- **WeeklyHydro_R.csv** (for reference only)

Output files:

- **groupAlteck2016_R** (line 256, use in Shiny App, and for BEACH data)
- **fluxAlteck2016_R.csv** 
- 

## Required R-packages:

```{r, message=FALSE}

# Date-time functions

library("ggplot2")
library("chron")
library("stringr")
library("plyr")
library("dplyr")


```

## Working directory

```{r, message=FALSE}


getwd()

```

## Import data

```{r, message=FALSE}

grpAlteck = read.csv2("Data/hydroAlteck2016_R.csv")
head(grpAlteck)

grpAlteck$Date = as.POSIXct(strptime(grpAlteck$Date, "%Y-%m-%d %H:%M", tz="EST"))
class(grpAlteck$Date)
sum(is.na(grpAlteck$Date))
sum(is.na(grpAlteck$Q.HW1))

```

## Define the Weekly discharge tags

```{r, tidy=TRUE}

grpAlteck$SubWeeks = NA
grpAlteck$SubWeeks[grpAlteck$Date< as.POSIXct('2016-03-25 12:04:00', tz="EST")]=as.character('W0-0x') # Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-03-25 12:04:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-03-28 22:37:00' , tz="EST")] = as.character('W0-1')

#
grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-03-28 22:37:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-03-30 12:17:00' , tz="EST")] = as.character('W0-2x') # Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-03-30 12:17:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-03-31 15:35:00' , tz="EST")] = as.character('W1-1')  

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-03-31 15:35:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-04-01 14:55:00' , tz="EST")] = as.character('W1-2')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-01 14:45:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-04-05 15:07:00' , tz="EST")] = as.character('W1-3x') # Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-05 15:07:00' , tz="EST") 
                         & grpAlteck$Date < as.POSIXct('2016-04-06 14:51:00' , tz="EST")] = as.character('W2-1')  

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-06 14:51:00' , tz="EST") 
                         & grpAlteck$Date < as.POSIXct('2016-04-09 00:38:50')] = as.character('W2-2')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-08 00:38:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-04-14 13:51:00' , tz="EST")] = as.character('W2-3x') # Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-14 13:51:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-04-16 18:32:00' , tz="EST")] = as.character('W3-1')  

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-16 18:32:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-04-17 09:02:00' , tz="EST")] = as.character('W3-2')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-17 09:02:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-04-18 20:30:00' , tz="EST")] = as.character('W3-2.1x') # Not smapled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-18 20:30:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-04-21 09:11:00' , tz="EST")] = as.character('W3-3')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-21 09:11:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-04-23 06:37:00' , tz="EST")] = as.character('W4-1')

#
grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-23 06:37:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-04-26 11:50:00' , tz="EST")] = as.character('W4-2x') # Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-04-26 11:50:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-01 10:46:00' , tz="EST")] = as.character('W5-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-01 10:46:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-03 12:02:00' , tz="EST")] = as.character('W5-2')

#
grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-03 12:02:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-03 13:09:00' , tz="EST")] = as.character('W5-3x')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-03 13:09:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-10 00:05:00' , tz="EST")] = as.character('W6-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-10 00:05:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-12 06:33:00' , tz="EST")] = as.character('W6-2')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-12 06:33:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-12 09:12:00' , tz="EST")] = as.character('W6-3')


grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-12 09:12:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-12 12:52:00' , tz="EST")] = as.character('W6-4')

#
grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-12 12:52:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-13 12:05:00' , tz="EST")] = as.character('W6-5x') # Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-13 12:05:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-16 15:11:00' , tz="EST")] = as.character('W7-1')

#
grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-16 15:11:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-17 09:16:00' , tz="EST")] = as.character('W7-2x') # Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-17 09:16:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-23 18:02:00' , tz="EST")] = as.character('W8-1')

#
grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-23 18:02:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-24 12:00:00' , tz="EST")] = as.character('W8-2x') # Not sampled


grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-24 12:00:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-29 12:09:00' , tz="EST")] = as.character('W9-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-29 12:09:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-30 05:48:00' , tz="EST")] = as.character('W9-2')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-30 05:48:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-30 12:11:00' , tz="EST")] = as.character('W9-3')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-30 12:11:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-30 17:28:00' , tz="EST")] = as.character('W9-4')
##
grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-30 17:28:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-05-31 12:00:00' , tz="EST")] = as.character('W9-5x') # Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-05-31 12:00:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-02 12:57:00' , tz="EST")] = as.character('W10-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-02 12:57:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-03 12:05:00' , tz="EST")] = as.character('W10-2')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-03 12:05:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-04 08:35:00' , tz="EST")] = as.character('W10-3')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-04 08:35:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-04 11:00:00' , tz="EST")] = as.character('W10-4')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-04 11:00:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-04 15:31:00' , tz="EST")] = as.character('W10-5')

# 
grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-04 15:31:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-07 12:00:00' , tz="EST")] = as.character('W10-6x') # Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-07 12:00:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-10 05:25:00' , tz="EST")] = as.character('W11-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-10 05:25:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-14 12:34:00' , tz="EST")] = as.character('W11-2')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-14 12:34:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-14 13:06:00' , tz="EST")] = as.character('W11-3')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-14 13:06:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-15 08:14:00' , tz="EST")] = as.character('W12-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-15 08:14:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-16 08:21:00' , tz="EST")] = as.character('W12-2')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-16 08:21:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-17 00:49:00' , tz="EST")] = as.character('W12-3')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-17 00:49:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-17 11:05:00' , tz="EST")] = as.character('W12-4')

#
grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-17 11:05:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-21 12:00:00' , tz="EST")] = as.character('W12-5x')# Not sampled

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-21 12:00:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-24 14:51:00' , tz="EST")] = as.character('W13-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-24 14:51:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-25 07:49:00' , tz="EST")] = as.character('W13-2')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-25 07:49:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-06-28 08:55:00' , tz="EST")] = as.character('W13-3')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-06-28 08:55:00' , tz="EST") & 
                           grpAlteck$Date < as.POSIXct('2016-07-04 14:41:00' , tz="EST")] = as.character('W14-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-07-04 14:41:00', tz="EST") & 
                           grpAlteck$Date <= as.POSIXct('2016-07-12 10:20:00', tz="EST")] = as.character('W15-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-09-08 00:00:00', tz="EST") & 
                           grpAlteck$Date <= as.POSIXct('2016-09-08 23:59:00', tz="EST")] = as.character('W16-1')

grpAlteck$SubWeeks[grpAlteck$Date >= as.POSIXct('2016-09-30 00:00:00', tz="EST") & 
                           grpAlteck$Date <= as.POSIXct('2016-09-30 23:59:00', tz="EST")] = as.character('W17-1')

head(grpAlteck)
sum(is.na(grpAlteck$Q.m3Hrs))
sum(is.na(grpAlteck$SubWeeks))
sum(is.na(grpAlteck$Q.HW1))

```

## Define new sub-IDs 

```{r}
Split <- strsplit(grpAlteck$SubWeeks, "-", fixed = TRUE)
grpAlteck$Weeks <- sapply(Split, "[", 1)

Split2 <- strsplit(grpAlteck$SubWeeks, "W", fixed = TRUE)
grpAlteck$WeekNo <- sapply(Split2, "[", 2)

Split3 <- strsplit(grpAlteck$WeekNo, "-", fixed=T)
grpAlteck$WeekNo <- sapply(Split3, "[", 1)
grpAlteck$WeekNo = as.numeric(grpAlteck$WeekNo)




head(grpAlteck)

```

## Classifying events

```{r}

library(plyr)
library(dplyr)
library(gridExtra)
library("scales")

detach("package:plyr", unload=TRUE) 
Q1change <- mutate(grpAlteck, Row = 1:n()) %>%
  mutate(PercentChange = (Q.HW1-lag(Q.HW1))/lag(Q.HW1) * 100)
Q1change$PercentChange[1] <- 0

dd <- Q1change[, c("Date", "Q.HW1")]
# set the number of rows to reduce by
idx <- ceiling(seq_len(nrow(Q1change)) / 10)
# do colMeans on the last column with lapply
# bind them with rbind to give a matrix, then convert to as.data.frame
res <- as.data.frame(do.call(rbind, lapply(split(dd[ncol(dd)], idx), 
                 colMeans, na.rm = TRUE)))
# assign first value of "Date" in each n-th group to the new dataframe
res$Date <- dd$Date[seq(1, nrow(dd), by=10)]

# Compute the %change on every row
res1 <- mutate(res, Event = 1:n()) %>%
  mutate(PercentChange = (Q.HW1-lag(Q.HW1))/lag(Q.HW1) * 100)

res1 <- mutate(res1, Event = 1:n()) %>%
  mutate(Change = (Q.HW1-lag(Q.HW1)))

res1$PercentChange[1] <- 0

res1$Markers1 <- ifelse(  res1$Change > 5 , res1$Q.HW1, NA)
res1$Markers2 <- ifelse(  res1$PercentChange > 90 & res1$Q.HW1 >20 & res1$Q.HW1 < 100, res1$Q.HW1, NA)

res1$Markers <- ifelse(!is.na(res1$Markers1), res1$Markers1, 
                       ifelse(!is.na(res1$Markers2), res1$Markers2, NA))
res <- res1[complete.cases(res1["Markers"]),]

resTime <- mutate(res, Event = 1:n()) %>%
  mutate(TimeDiff = Date-lag(Date))

resTime$TimeDiff[1]<-1440
resTime <- resTime[resTime$TimeDiff > 20, ]

# For some reason, changed minutes to hrs
resTime <- mutate(resTime, Event = 1:n()) %>%
  mutate(TimeDiff = Date-lag(Date))

# Add first row time, so as to not loose it
resTime$TimeDiff[1]<-24
resTime <- resTime[resTime$TimeDiff > 5, ]
resTime <- mutate(resTime, Event = 1:n()) %>%
  mutate(TimeDiff = Date-lag(Date))

resTime$TimeDiff[1]<-24
resTime <- resTime[resTime$TimeDiff >= 9, ]
resTime <- mutate(resTime, Event = 1:n()) %>%
  mutate(TimeDiff = Date-lag(Date))

resTime$TimeDiff[1]<-24
resTime <- resTime[resTime$TimeDiff > 12, ]
resTime <- mutate(resTime, Event = 1:n()) %>%
  mutate(TimeDiff = Date-lag(Date))

resTime$TimeDiff[1]<-24
resTime$Markers1 <- NULL
resTime$Markers2 <- NULL
resTime$Q.HW1 <- NULL

grpAlteck <- merge(grpAlteck, resTime, by= "Date", all = T)
sum(is.na(grpAlteck$Q.HW1))

grpAlteck$numNoEvent <- ifelse(grpAlteck$Q.HW1 < 18, 1, 0)
grpAlteck$numNoEvent <- ifelse(is.na(grpAlteck$numNoEvent), 0, grpAlteck$numNoEvent)
sum(is.na(grpAlteck$numNoEvent))

# No Event counts the number of hours without a major discharge (i.e. 18 m3/h)
noEventVect <- grpAlteck$numNoEvent
aux <- split(noEventVect, cumsum(noEventVect == 0))
v <- unlist(sapply(aux, cumsum))
grpAlteck$noEventHrs <- v/60 # Convert to Hrs

# No rain counts for Hrs without at least 0.2 mm rain (in a 2 minute interval)
grpAlteck$numNoRain <- ifelse(grpAlteck$Rain.mm < 0.2, 1, 0)
grpAlteck$numNoRain <- ifelse(is.na(grpAlteck$numNoRain), 1, grpAlteck$numNoRain)
sum(is.na(grpAlteck$numNoRain))

noRainVect <- grpAlteck$numNoRain
aux2 <- split(noRainVect, cumsum(noRainVect == 0))
v2 <- unlist(sapply(aux2, cumsum))
grpAlteck$dryhrs <- v2/60

#length(grpAlteck$numNoEvent)
#length(noEventVect)

#grpAlteck$minSinceEvent <- NA
#cumDuration <- 0
# if Q.HW1 < 10m3h: cumDuration += 3 min
# else cumDuration = 0
# for every cell in Date, assign cumDuration to timeSinceEvent
#for (i in 1:length(grpAlteck$Date)) {
#  if (grpAlteck[i,]['Q.HW1'] < 17){
#    grpAlteck[i,]['minSinceEvent'] = cumDuration
#    cumDuration <- cumDuration + 3.0
#  } else {
#    cumDuration <- 0
#    grpAlteck[i,]['minSinceEvent'] = cumDuration
#  }
#}


# (Sub)Event markers
ggplot() +
  geom_line(data = grpAlteck, aes(x= Date, y = Q.HW1), color = "blue") +
  geom_point(data = grpAlteck, aes(x= Date, y = Markers), color = "forestgreen") +
  scale_y_continuous(trans=log_trans(), breaks=c(1,5,8, 10,25, 50, 80, 100,1000))

ggplot() +
  geom_line(data = grpAlteck, aes(x= Date, y = dryhrs), color = "blue")
  
```




Save the file in current state, as it is needed in the App. 

```{r}
write.csv2(grpAlteck, "Data/groupAlteck2016_R.csv", row.names = F)
```



## Characterize discharge sub-weeks (i.e. sampled discharge) 

The data frame produced will include, for each sub-week, the:

- initial time (ti)
- final time (tf)
- initial discharge (iflux)
- final discharge (fflux)
- change in discharge between ti and tf (changeflux)
- change in discharge to extreme (maxQ or minQ) withing subsample (chExtreme)
- discharge at maxQ (maxQ)
- minimum discharge (minQ)
- elapsed time in hours (tdiff)

```{r}
colnames(grpAlteck)

dflux = grpAlteck %>% 
  group_by(SubWeeks) %>% 
  # filter(Type == 'Sample') %>%
  select(Date, Q.HW1, SubWeeks, noEventHrs, dryhrs) %>%
  summarise(ti= Date[1], 
            tf = Date[length(Date)], 
            iflux = Q.HW1[1], 
            fflux = Q.HW1[length(Q.HW1)], 
            changeflux = (Q.HW1[length(Q.HW1)] - Q.HW1[1]), 
            maxQ = max(Q.HW1), 
            minQ = min(Q.HW1)  ,
            dryHrsIni = dryhrs[1],
            dryHrsMax = max(dryhrs),
            dryHrsAve = mean(dryhrs),
            noEventHrsIni = noEventHrs[1],
            noEventHrsMax = max(noEventHrs),
            noEventHrsAve = mean(noEventHrs)
            )



# Time elapsed within sub-week in hrs (sampled and non-sampled)
dflux$Duration.Hrs = 
  as.numeric(difftime(dflux$tf, dflux$ti, units = "hours"), units = "hours")

# "chngeExtreme" is computed as:
# If change in flux within subsample is: 
# negative, maxQminQ =  (min. discharge) - (initial discharge)
# positive, maxQminQ =  (max. discharge) - (initial discharge)
dflux$chExtreme <- NA

dflux$chExtreme[dflux$changeflux <= 0] = 
  dflux$minQ[dflux$changeflux <= 0] - dflux$iflux[dflux$changeflux <= 0]

dflux$chExtreme[dflux$changeflux > 0] = 
  dflux$maxQ[dflux$changeflux > 0] - dflux$iflux[dflux$changeflux > 0]

EventMarker <- grpAlteck[, c("SubWeeks", "Event", "Markers", "TimeDiff")]
EventMarker <- EventMarker[!is.na(EventMarker$Event),]

dflux <- merge(dflux, EventMarker, by = "SubWeeks", all = T)

colnames(dflux)[1] <- "WeekSubWeek"
head(dflux)
```


## Saving

```{r}

write.csv2(dflux, "Data/fluxAlteck2016_R.csv", row.names = F)
sum(is.na(dflux$maxQ))

```
