==> 06_MassSoils_Composite.Rmd <==
---
title: "Mass Soils - Composite Weeks Alteck 2016"
author: "PAZ"
date: "November 2016"
output: 
  pdf_document:
    fig_caption: yes
---


```{r, echo=FALSE, message=FALSE, include=FALSE}
Sys.setlocale("LC_ALL", "English")
```

## Purpose

This file merges weekly composite concentrations and isotope data. 

Imports: 

- **SoilCompConc_W1toW15.csv** 
- **SoilCompIsotopes_W1toW15.csv** 

Generates:

- **WeeklySoils_Rng.csv**

## Required R-packages:

```{r, message=FALSE}

library("plyr")
library("dplyr")

```

## Working directory

```{r, message=FALSE}

# setwd("D:/Documents/these_pablo/Alteckendorf2016/R")
# setwd("/Users/DayTightChunks/Documents/PhD/Routput/Alteck/R")
# setwd("D:/Documents/these_pablo/Alteckendorf2016/00_TransparencyFolder")
getwd()

```

## Lab Parameters

```{r}
source("global.R")
```

## Field Assumptions

```{r}
# S-metolachlor Mass [g]
# Conc. [ug/g dry soil] * [g/10^6 ug] * density [g/m3] * depth [m]* A [m2] 
# Soil bulk density: 2.2 [g/cm3] or 0.99 [g/cm3] -> Leaching experiments: 0.99 [g/cm3] or 0.99 x 10^6 g/m3

# Fresh tilled
rho1 = 0.98*10^6 # soil density [g/m3] Leaching experiment, "freshly tilled" soil
depth1 = 0.015 # [m] # Fresh tilled

rho1.Sa = 0.98*10^6 # S is different due to less area prop being fresh tilled 
depth1.Sa = 0.011 # [m]

rho1.Sb = 0.98*10^6 # S is different due to fresh tilling on Burger near April 14 
depth1.Sb = 0.015 # [m] 

rho1.T = 0.98*10^6 # T is different due to observed early crusting 
depth1.T = 0.01 # [m] 

# Mid April
# rho2 = 2.20*10^6 # soil density [g/m3] Lefrancq's measuremnts after major May rainfall event
rho2 = 0.99*10^6
depth2 = 0.015 # [m] # Mid april

# Early May
rho3 = 0.99*10^6 # 0.99*10^6
depth3 = 0.0046 # [m] # Early may

# Extrapolated Transect Areas pre-corn applications
# Area_Na = 13.92663*10^4 # [m2]

# Extrapolated
# Assumptions on Non-measured plots - Transect assignment:
#Area_Ta = 6.55813*10^4  # [m2] # South Burger's as Talweg
#Area_Sa = 11.05376*10^4 # [m2] # South Burger's as Talweg
# Area_Ta = 4.37134*10^4  # [m2]
# Area_Sa = 13.3175*10^4 # [m2] # South Burger's as South



# Transect Areas post Corn applications (not on transect)
#Area_Nb = 14.9949*10^4 # [m2] 

# Assumptions on Non-measured plots - Transect assignment:
#Area_Tb = 6.55813*10^4  # [m2] # South Burger's as Talweg
#Area_Sb = 11.65202*10^4 # [m2] # South Burger's as Talweg
#Area_Tb = 4.37134*10^4  # [m2]
#Area_Sb = 13.91767*10^4 # [m2] # South Burger's as South

# Not extrapolated (only where mass was applied & measured)
# Area_Na = 84580.901
# Area_Ta = 29863.690
# Area_Sa = 94205.501

Area_Nb = 101721.702
Area_Tb = 39247.330
Area_Sb = 100245.721

Area_Na = Area_Nb
Area_Ta = Area_Tb
Area_Sa = Area_Sb

Area_tot <- Area_Nb + Area_Tb + Area_Sb
Area_tot/10^4 
  
```


## Composite Concentrations & Isotope Data - Alteckendorf 2016

1. Import CSV files 

```{r, message=FALSE}

#weeklySoilConc = read.csv2("Data/SoilCompConc_W1toW15.csv", header = TRUE)
# Date format stopped working in CSV
# Convert in CSV via  "=TEXT(CELL.ID, "dd/mm/yyyy hh:mm")" based on xls-file date

weeklySoilConc = read.csv2("Data/SoilCompConc_W1toW15.csv", header = TRUE, dec = ".")
if (length(weeklySoilConc) == 1){
  weeklySoilConc = read.csv("Data/SoilCompConc_W1toW15.csv", header = TRUE)
}
head(weeklySoilConc)

weeklySoilConc$Date.ti <- as.POSIXct(strptime(weeklySoilConc$Date.Soil, "%d/%m/%Y %H:%M", tz="EST")) 
sum(is.na(weeklySoilConc$Date.ti))
# View(weeklySoilConc)
names(weeklySoilConc)
weeklySoilConc <- weeklySoilConc[,c("Filename",
                                    "Transect",
                                    "Wnum",
                                    "Date.Soil",
                                    "Date.ti",
                                    "Conc.mug.g.dry.soil",
                                    "Conc.ComSoil.SD",
                                     "Mass.Soil.g", "theta.prct")]

colnames(weeklySoilConc)[colnames(weeklySoilConc) == "Filename"] <- "ID"
print("Soil Composites- Concentrations")
str(weeklySoilConc)

# After nanogram revision
weeklySoilIso  = read.csv2("Data/SoilCompIsotopes_W1toW15.csv", header = TRUE, dec = ".")
if (length(weeklySoilIso) == 1){
  weeklySoilIso = read.csv("Data/SoilCompIsotopes_W1toW15.csv", header = T)
}
colnames(weeklySoilIso)
weeklySoilIso$ID <- NULL
weeklySoilIso <- weeklySoilIso[complete.cases(weeklySoilIso[, "d.13C.12C"]), ]

# Correct for soil shift
weeklySoilIso$d.13C.12C = round( (weeklySoilIso$d.13C.12C - meanshift), 1) 
weeklySoilIso$DD13 <- weeklySoilIso$d.13C.12C - initialDelta
colnames(weeklySoilIso)[colnames(weeklySoilIso) == "Filename"] <- "ID"

weeklySoilIso <- weeklySoilIso[, c("ID",
                                   # "Repl",
                                   "d.13C.12C",
                                   "DD13"
                                   )]

isoCompSummary = ddply(weeklySoilIso, c("ID"), summarise,
                         N_compsoil    = length(d.13C.12C),
                         comp.d13C = mean(d.13C.12C),
                         comp.d13C.SD = sd(d.13C.12C),
                         N_isoComp = length(d.13C.12C)
                         # comp.d13C.SE = comp.d13C.SD / sqrt(N_compsoil),
                         # N_ngC = length(ngC),
                         # ngC.mean = mean(ngC),
                         # ngC.SD = sd(ngC) #,
                         # ngC.SE = ngC.SD/sqrt(N_ngC)
                       )

isoCompSummary$prctError <- (isoCompSummary$comp.d13C.SD/isoCompSummary$comp.d13C)*-100
mean(!is.na(isoCompSummary$prctError))

# sum(isoCompSummary$N_ngC == 2)/(sum(isoCompSummary$N_ngC == 2) + sum(isoCompSummary$N_ngC > 2))


print("Soil Composites - Isotopes All")
str(weeklySoilIso)

print("Soil Composites - Isotopes Ave and St.Dev.")
str(isoCompSummary)
```

2. Merge lab concentrations and isotopes

```{r}
comp.CoIs = merge(weeklySoilConc, isoCompSummary, by = "ID", all = T)
comp.CoIs$Wnum = as.numeric(comp.CoIs$Wnum)
comp.CoIs <- comp.CoIs[order(comp.CoIs$Wnum),]
comp.CoIs$DD13C.comp <- (comp.CoIs$comp.d13C - (d13Co))

comp.CoIs$comp.IMP.d13C <- comp.CoIs$comp.d13C
comp.CoIs$comp.IMP.d13C[is.na(comp.CoIs$comp.d13C)] <- 
  ave(comp.CoIs$comp.d13C,
      comp.CoIs$Wnum,
      FUN= function(x) mean(x, na.rm = T))[is.na(comp.CoIs$comp.d13C)] 


# comp.CoIs$comp.d13C <- ifelse(is.na(comp.CoIs$comp.d13C), comp.CoIs$comp.IMP.d13C, comp.CoIs$comp.d13C) 

print("Merged Soil Concentrations and Isotopes")
str(comp.CoIs)  

write.csv2(comp.CoIs, 
           'Data/MonitoringScopeSoils_R.csv', row.names = F)
```


3. Compute Soil S-metolachlor Mass at time $t$ across space  

For non-measured plots, the soil concentration and isotope measured at the neareast transect is assumed. The total area for each transect at the end of the season is shown below. Corn fields in the catchment were known to have received S-metolachlor applications only during the last week of May, 2017. Given that two of these plots were not included within the transects, their area was not accounted for but until after the known application dates for corn plots. 

![Transect Areas $[Ha]$ (North: 14.995; Talweg: 4.371; South: 13.918)](Images/TransectAreas.png)

```{r}
# Check values:
Area_Nb/10000 
Area_Tb/10000 
Area_Sb/10000

```


The total pesticide mass for each transect at time $t$ is then given by:

$$
M(t)_{Ta} = C(t)_{T} \cdot \rho \cdot A_T \cdot D  
$$



```{r, echo=T}

# S-metolachlor Mass [g]
# Conc. [ug/g dry soil] * [g/10^6 ug] * density [g/m3] * depth [m]* A [m2] 
# Soil bulk density: 2200 or 0.99? -> Leaching experiments: 0.99 [g/cm3]
rho1 # soil density [g/m3]
rho2 # soil density [g/m3]
depth1 # [m]
depth2

# Transect Areas pre-corn applications
Area_Na # [m2]
Area_Ta  # [m2]
Area_Sa # [m2]
Area_early <- Area_Na + Area_Ta + Area_Sa
# Transect Areas post Corn applications (not on transect)
Area_Nb # [m2] 
Area_Tb # [m2]
Area_Sb # [m2]

# Assign new column for S-metolachlor mass in grams
comp.CoIs$MassSoil.g <- NA
comp.CoIs$MassSoil.g.SD <- NA

# Areas with S-metolachlor before week 9 (no longer used, as Effected Composite Area is used)
#comp.CoIs$MassSoil.g <- 
#  ifelse((comp.CoIs$Transect == "N" & comp.CoIs$Wnum < 8), 
#         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho*depth*Area_Na,
#  ifelse((comp.CoIs$Transect == "T" & comp.CoIs$Wnum < 8), 
#         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho*depth*Area_Ta, 
#  ifelse((comp.CoIs$Transect == "S" & comp.CoIs$Wnum < 8), 
#         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho*depth*Area_Sa, comp.CoIs$MassSoil.g)))

#comp.CoIs$MassSoil.g.SD <- 
#  ifelse((comp.CoIs$Transect == "N" & comp.CoIs$Wnum < 8), 
#         comp.CoIs$Conc.ComSoil.SD*10^-6*rho*depth*Area_Na,
#  ifelse((comp.CoIs$Transect == "T" & comp.CoIs$Wnum < 8), 
#         comp.CoIs$Conc.ComSoil.SD*10^-6*rho*depth*Area_Ta, 
#  ifelse((comp.CoIs$Transect == "S" & comp.CoIs$Wnum < 8), 
#         comp.CoIs$Conc.ComSoil.SD*10^-6*rho*depth*Area_Sa, comp.CoIs$MassSoil.g.SD)))

# Difference in Bulk Density after 100 mm cumulative rain
comp.CoIs$MassSoil.g <- 
  ifelse((comp.CoIs$Transect == "N" 
          & comp.CoIs$Date.ti <= as.POSIXct("2016-04-14 13:52:00", tz = "EST")), 
         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho1*depth1*Area_Na,
  ifelse((comp.CoIs$Transect == "T" 
          & comp.CoIs$Date.ti <= as.POSIXct("2016-04-14 13:52:00", tz = "EST")), 
         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho1.T*depth1.T*Area_Ta, 
  ifelse((comp.CoIs$Transect == "S" 
          & comp.CoIs$Date.ti < as.POSIXct("2016-04-14 13:52:00", tz = "EST")), # Note only "<""
         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho1.Sa*depth1.Sa*Area_Sa, comp.CoIs$MassSoil.g)))

comp.CoIs$MassSoil.g.SD <- 
  ifelse((comp.CoIs$Transect == "N" & comp.CoIs$Date.ti <= as.POSIXct("2016-04-14 13:52:00", tz = "EST")), 
         comp.CoIs$Conc.ComSoil.SD*10^-6*rho1*depth1*Area_Na,
  ifelse((comp.CoIs$Transect == "T" & comp.CoIs$Date.ti <= as.POSIXct("2016-04-14 13:52:00", tz = "EST")), 
         comp.CoIs$Conc.ComSoil.SD*10^-6*rho1*depth1*Area_Ta, 
  ifelse((comp.CoIs$Transect == "S" & comp.CoIs$Date.ti < as.POSIXct("2016-04-14 13:52:00", tz = "EST")), 
         comp.CoIs$Conc.ComSoil.SD*10^-6*rho1*depth1*Area_Sa, comp.CoIs$MassSoil.g.SD)))

# Mid april - Mid May
comp.CoIs$MassSoil.g <-
  ifelse((comp.CoIs$Transect == "N" 
          & comp.CoIs$Date.ti > as.POSIXct("2016-04-14 13:52:00", tz = "EST")
          & comp.CoIs$Date.ti <= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho2*depth2*Area_Na,
       ifelse((comp.CoIs$Transect == "T" 
          & comp.CoIs$Date.ti > as.POSIXct("2016-04-14 13:52:00", tz = "EST")
          & comp.CoIs$Date.ti <= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho2*depth2*Area_Ta, 
       ifelse((comp.CoIs$Transect == "S" 
               & comp.CoIs$Date.ti == as.POSIXct("2016-04-14 13:52:00", tz = "EST")), 
              comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho1.Sb*depth1.Sb*Area_Sa,
          ifelse((comp.CoIs$Transect == "S" 
            & comp.CoIs$Date.ti > as.POSIXct("2016-04-14 13:52:00", tz = "EST")
            & comp.CoIs$Date.ti <= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
            comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho2*depth2*Area_Sa, comp.CoIs$MassSoil.g))))

comp.CoIs$MassSoil.g.SD <- 
  ifelse((comp.CoIs$Transect == "N" 
          & comp.CoIs$Date.ti > as.POSIXct("2016-04-14 13:52:00", tz = "EST")
          & comp.CoIs$Date.ti <= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.ComSoil.SD*10^-6*rho2*depth2*Area_Na,
  ifelse((comp.CoIs$Transect == "T" 
          & comp.CoIs$Date.ti > as.POSIXct("2016-04-14 13:52:00", tz = "EST")
          & comp.CoIs$Date.ti <= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.ComSoil.SD*10^-6*rho2*depth2*Area_Ta, 
  ifelse((comp.CoIs$Transect == "S" 
          & comp.CoIs$Date.ti > as.POSIXct("2016-04-14 13:52:00", tz = "EST")
          & comp.CoIs$Date.ti <= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.ComSoil.SD*10^-6*rho2*depth2*Area_Sa, comp.CoIs$MassSoil.g.SD)))

# Mid May onwards
comp.CoIs$MassSoil.g <- 
  ifelse((comp.CoIs$Transect == "N" & comp.CoIs$Date.ti > as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho3*depth3*Area_Nb,
  ifelse((comp.CoIs$Transect == "T" & comp.CoIs$Date.ti >= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho3*depth3*Area_Tb, 
  ifelse((comp.CoIs$Transect == "S" & comp.CoIs$Date.ti >= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.mug.g.dry.soil*10^-6*rho3*depth3*Area_Sb, comp.CoIs$MassSoil.g)))

comp.CoIs$MassSoil.g.SD <- 
  ifelse((comp.CoIs$Transect == "N" & comp.CoIs$Date.ti >= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.ComSoil.SD*10^-6*rho3*depth3*Area_Nb,
  ifelse((comp.CoIs$Transect == "T" & comp.CoIs$Date.ti >= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.ComSoil.SD*10^-6*rho3*depth3*Area_Tb, 
  ifelse((comp.CoIs$Transect == "S" & comp.CoIs$Date.ti >= as.POSIXct("2016-05-10 00:06:00", tz = "EST")), 
         comp.CoIs$Conc.ComSoil.SD*10^-6*rho3*depth3*Area_Sb, comp.CoIs$MassSoil.g.SD)))

# Areas as variables (for later computation of bulk catchment mass)
comp.CoIs$Area.N <-
  ifelse((comp.CoIs$Wnum < 8), Area_Na, Area_Nb)
         
comp.CoIs$Area.T <-
  ifelse((comp.CoIs$Wnum < 8), Area_Ta, Area_Tb)

comp.CoIs$Area.S <-
  ifelse((comp.CoIs$Wnum < 8), Area_Sa, Area_Sb)
  
# Needed to compute Transect res
#comp.CoIs$Area_Nt <- Area_Nt  # [m2] 
#comp.CoIs$Area_Tt <- Area_Tt 
#comp.CoIs$Area_St <- Area_St

names(comp.CoIs)
print("S-meto mass per transect at time-t")
str(comp.CoIs)
tail(comp.CoIs)

write.csv2(comp.CoIs, 'Data/WeeklySoils_Rng.csv', row.names = F)
```
