==> 07_Soils&DischMerged.Rmd <==
---
title: "Soils & Discharge Merged"
author: "PAZ"
date: "17 novembre 2016"
output: pdf_document
---

```{r, echo=FALSE, message=FALSE, include=FALSE}
Sys.setlocale("LC_ALL", "English")
```

## Purpose

This file merges outlet data with soil data to plot cumulative exported and remaining S-metolachlor mass. The nearest soil sample date for each transect is used to match the initial time ("ti") of the sampling discharge period. This is most adequate merging location given that samples took place shortly before relaunching the automatic sampler. 

Note that week numbers for water and soils are offset by one. I.e. Week 1 soils influence/regulate Week 2's water sample results. 

Imports: 

- **WeeklyHydroContam_R.csv** (Book 5 - Mass Discharged Outlet)
- **WeeklySoils_Rng.csv** (Book 06 - Mass Soils Composite)

Generates:

- **WeekSoilHydroCont_R.csv**
- **MassBalance_R.csv**

## Required R-packages:

```{r, message=FALSE}

library("plyr")
library("dplyr")

```

## Working directory

```{r, message=FALSE}

# setwd("D:/Documents/these_pablo/Alteckendorf2016/R")
# setwd("/Users/DayTightChunks/Documents/PhD/Routput/Alteck/R")
# setwd("D:/Documents/these_pablo/Alteckendorf2016/00_TransparencyFolder")
getwd()

```

# Lab parameters

```{r}
source("global.R")
```


## Import files

```{r}
outlet = read.csv2("Data/WeeklyHydroContam_R.csv", header = T)
outlet$ti <- as.POSIXct(outlet$ti, "%Y-%m-%d %H:%M", tz = "EST")
sum(is.na(outlet$ti))


names(outlet)
# Select variables from Water dataset
outlet <- outlet[, c("ti", "WeekSubWeek", "Event", "Duration.Hrs", 
                     "timeSinceApp", "timeSinceApp.NoSo",
                     "timeSinceApp.N",  "timeSinceApp.T",  "timeSinceApp.S",
                     "diss.d13C", "SD.d13C",
                     "CumOutDiss.g", "CumOutFilt.g", 
                     "TotSMout.g", "TotSMout.g.SD" , 
                     "MELsm.g", "MELsm.g.SD",
                     "Appl.Mass.g" , "Appl.Mass.g.OT",
                     "CumAppMass.g", "CumAppMass.g.N", "CumAppMass.g.T", "CumAppMass.g.S", 
                     "CumAppMass.g.OT", "CumAppMass.g.N.OT", "CumAppMass.g.T.OT", "CumAppMass.g.S.OT",
                     "iniCo.ug.g.N",  "iniCo.ug.g.T", "iniCo.ug.g.S",  
                     "CumOutSmeto.g", "CumOutMELsm.g")]
print("Water")
names(outlet)

soils = read.csv2("Data/WeeklySoils_Rng.csv", header =T) # Corrected with only ngC > 2.0 
names(soils)
soils$Date.ti <- as.POSIXct(soils$Date.ti, "%Y-%m-%d %H:%M", tz = "EST")
#soils$Date.ti <- as.POSIXct(soils$Date.ti, "%d/%m/%Y %H:%M", tz = "EST")
sum(is.na(soils$Date.ti))

print("Soils")
str(soils)
```


# Get soil concentrations for each transect for merging horizontally

```{r}

# North
soils.N <- subset(soils, soils$Transect == "N")
soils.N <- soils.N[, c("Date.ti",
                       "MassSoil.g", "MassSoil.g.SD", 
                       "Conc.mug.g.dry.soil", 
                       "comp.d13C", "comp.d13C.SD", # "comp.d13C.SE", 
                       "ID", 
                       "Area.N", "Area.T", "Area.S" #,
                       # "Area_Nt", "Area_Tt", "Area_St"
                       )]

colnames(soils.N) <- c("ti",
                       "MassSoil.g.North", "MassSoil.g.SD.North",
                       "Conc.mug.g.dry.soil.N",
                       "comp.d13C.North", "comp.d13C.SD.North", # "comp.d13C.SE.North",  
                       "ID.N", 
                       "Area.N", "Area.T", "Area.S"#,
                       #"Area.Nt", "Area.Tt", "Area.St"
                       )

# Talweg
soils.T <- subset(soils, soils$Transect == "T")
soils.T <- soils.T[, c("Date.ti",
                       "MassSoil.g", "MassSoil.g.SD",
                       "Conc.mug.g.dry.soil",
                       "comp.d13C", "comp.d13C.SD", # "comp.d13C.SE", 
                       "ID")]

colnames(soils.T) <- c("ti",
                       "MassSoil.g.Talweg", "MassSoil.g.SD.Talweg",
                       "Conc.mug.g.dry.soil.T",
                       "comp.d13C.Talweg", "comp.d13C.SD.Talweg", # "comp.d13C.SE.Talweg", 
                       "ID.T" )

# South
soils.S <- subset(soils, soils$Transect == "S")
soils.S <- soils.S[, c("Date.ti", 
                       "MassSoil.g", "MassSoil.g.SD",
                       "Conc.mug.g.dry.soil",
                       "comp.d13C", "comp.d13C.SD", # "comp.d13C.SE", 
                       "ID" )]
colnames(soils.S) <- c("ti",  
                       "MassSoil.g.South", "MassSoil.g.SD.South",
                       "Conc.mug.g.dry.soil.S",
                       "comp.d13C.South", "comp.d13C.SD.South", # "comp.d13C.SE.South",
                       "ID.S" )

```

# Total Catchment Mass, Bulk Mass and Bulk Isotopes

Bulk isotopes are calculated based on the following:

$$
\delta_{bulk} =  \frac{ M_{North} }{ M_{tot} } \delta_{North} 
  + \frac{ M_{Talweg} }{ M_{tot} } \delta_{Talweg}
  + \frac{ M_{South} }{ M_{tot} } \delta_{South}
$$

Bulk mass (not currently used) is a proxy for the potential mass that can be discharged at a give time. It is calculated based on the pondered mass from each transect area and the proportion of that area in relation to the entire catchment such that:

$$
M_{bulk} =  \frac{ A_{North} }{ A_{tot} } M_{North} 
  + \frac{ A_{Talweg} }{ A_{tot} } M_{Talweg}
  + \frac{ A_{South} }{ A_{tot} } M_{South}
$$

$$
C_{bulk} =  \frac{ A_{North} }{ A_{tot} } C_{North} 
  + \frac{ A_{Talweg} }{ A_{tot} } C_{Talweg}
  + \frac{ A_{South} }{ A_{tot} } C_{South}
$$

```{r}

library(zoo)
class(outlet$ti)
class(soils.T$ti)

soilsOut <- merge(outlet, soils.N, by = "ti", all = T)
soilsOut <- merge(soilsOut, soils.T, by = "ti", all = T)
soilsOut <- merge(soilsOut, soils.S, by = "ti", all = T)

soilsOut$DD13C.North <- soilsOut$comp.d13C.North - initialDelta
soilsOut$DD13C.Talweg <- soilsOut$comp.d13C.Talweg - initialDelta
soilsOut$DD13C.South <- soilsOut$comp.d13C.South - initialDelta

# Total mass in catchment
soilsOut$CatchMassSoil.g <- 
  soilsOut$MassSoil.g.North + 
  soilsOut$MassSoil.g.Talweg + 
  soilsOut$MassSoil.g.South

soilsOut$CatchMassSoil.g.SD <- (
  ( soilsOut$MassSoil.g.SD.North^2 + 
      soilsOut$MassSoil.g.SD.South^2 + 
      soilsOut$MassSoil.g.SD.Talweg^2
    )/3
  )^0.5
  
# Approximate initial mass for first 6 observation (before recording composite soils)
#soilsOut$CatchMassSoil.g[1:6] <- ifelse(is.na(soilsOut$CatchMassSoil.g), 
#                                        soilsOut$CumAppMass.g-soilsOut$CumOutMELsm.g, 
#                                        soilsOut$CatchMassSoil.g) 


#soilsOut$BulkMass.g <- 
#  soilsOut$MassSoil.g.North*(soilsOut$Area.N/(soilsOut$Area.N+soilsOut$Area.T+soilsOut$Area.S)) +
#  soilsOut$MassSoil.g.Talweg*(soilsOut$Area.T/(soilsOut$Area.N+soilsOut$Area.T+soilsOut$Area.S)) +
#  soilsOut$MassSoil.g.South*(soilsOut$Area.S/(soilsOut$Area.N+soilsOut$Area.T+soilsOut$Area.S))

# Bulk catchment isotopes
soilsOut$BulkCatch.d13 <- 
  (soilsOut$MassSoil.g.North/soilsOut$CatchMassSoil.g)*soilsOut$comp.d13C.North + 
  (soilsOut$MassSoil.g.Talweg/soilsOut$CatchMassSoil.g)*soilsOut$comp.d13C.Talweg + 
  (soilsOut$MassSoil.g.South/soilsOut$CatchMassSoil.g)*soilsOut$comp.d13C.South

soilsOut$BulkCatch.d13.SD <- 
  (soilsOut$MassSoil.g.North/soilsOut$CatchMassSoil.g)*soilsOut$comp.d13C.SD.North + 
  (soilsOut$MassSoil.g.Talweg/soilsOut$CatchMassSoil.g)*soilsOut$comp.d13C.SD.Talweg + 
  (soilsOut$MassSoil.g.South/soilsOut$CatchMassSoil.g)*soilsOut$comp.d13C.SD.South

soilsOut$DD13.Bulk <-  (soilsOut$BulkCatch.d13 - (d13Co)) 

# Bulk catchment concentrations
soilsOut$Area.Catchment <- soilsOut$Area.N + soilsOut$Area.T + soilsOut$Area.S

soilsOut$BulkCatch.Conc <- 
  (soilsOut$Area.N/soilsOut$Area.Catchment)*soilsOut$Conc.mug.g.dry.soil.N + 
  (soilsOut$Area.T/soilsOut$Area.Catchment)*soilsOut$Conc.mug.g.dry.soil.T + 
  (soilsOut$Area.S/soilsOut$Area.Catchment)*soilsOut$Conc.mug.g.dry.soil.S

soilsOut$iniCo.Bulk <-
  soilsOut$iniCo.ug.g.N*(soilsOut$Area.N/soilsOut$Area.Catchment) +
  soilsOut$iniCo.ug.g.T*(soilsOut$Area.T/soilsOut$Area.Catchment) + 
  soilsOut$iniCo.ug.g.S*(soilsOut$Area.S/soilsOut$Area.Catchment)


```


```{r}

print("Merged Soils and Outlet by ti")
str(soilsOut)

write.csv2(soilsOut, 
           'Data/MassBalance_R.csv', row.names = F)
```



# Plot

```{r}
library("ggplot2")
library("scales")
library("reshape2")
library("cowplot")

# Melt data set
##Subset the necessary columns
soilsRemainMass <- soilsOut[, c("ti" ,"CumAppMass.g", "CumOutDiss.g", "CumOutFilt.g", "CumOutMELsm.g" , "CatchMassSoil.g")]
soilsRemainMass$CumAppMass.g[1:3] <- NA

# Replace Catchment Mass's NA with the most recent non-NA prior to it (assumes no degradation).
# Purpose: To match continuous outlet time array    
soilsRemainMass$CatchMassSoil.g <- na.locf(soilsRemainMass$CatchMassSoil.g)
# soilsRemainMass <- na.omit(soilsRemainMass)

##Then rearrange your data frame
remainMassMolten = melt(soilsRemainMass, id=c("ti"))

# View(remainMassMolten)

pg <- remainMassMolten

# Change variable names:
levels(pg$variable)[levels(pg$variable)=="CumAppMass.g"] <- "Applied SM Cum. (Survey)"
levels(pg$variable)[levels(pg$variable)=="CumOutMELsm.g"] <- "MEL-SM Cum. (Outlet)"
levels(pg$variable)[levels(pg$variable)=="CatchMassSoil.g"] <- "Remaining SM (1cm Soil)"

levels(pg$variable)[levels(pg$variable)=="CumOutDiss.g"] <- "Dissolved SM Cum. (Outlet)"
levels(pg$variable)[levels(pg$variable)=="CumOutFilt.g"] <- "Sediment SM Cum. (Outlet)"

# Change the order:
levels(pg$variable)
pg$variable <- factor(pg$variable, levels = c("Applied SM Cum. (Survey)",  "Remaining SM (1cm Soil)", "MEL-SM Cum. (Outlet)", "Dissolved SM Cum. (Outlet)", "Sediment SM Cum. (Outlet)" ))

pgSimple <- pg[which(pg$variable != ("Dissolved SM Cum. (Outlet)") & pg$variable != ("Sediment SM Cum. (Outlet)")), ]
# names(pg)[names(pg)=="variable"]  <- "Estimated Mass"

massBalTop <- ggplot(pg) + 
  geom_line(aes(x=ti, y=value, group = variable, color=variable)) +
  
  # Themes and axes
  theme_bw() +
  theme(# axis.text.x=element_text(angle = 45, hjust = 1), 
        axis.text.x=element_blank(),
        axis.title.x=element_blank(),
        legend.position="top"
        
        )+
  labs(color = "Estimated Mass") +
  guides(col = guide_legend(ncol = 1)) +  # Sets legend parameters
  
  # xlab("Date") +
  scale_x_datetime(breaks = date_breaks("2 weeks"), labels = date_format("%b %d")) +
  ylab(expression(paste("Mass ", {(g)}))) +
  # scale_y_continuous(breaks = c(100, 5000, 10000, 20000), limits = c(100, 20000) ) 
  scale_y_continuous(trans=log_trans(), breaks=c(1,5,10,50,100,500,1000,5000, 10000))
massBalTop


massBalBottom <- ggplot(pg) + 
  geom_line(aes(x=ti, y=value, color=variable)) +
  
  # Themes and axes
  theme_bw() +
  theme(axis.text.x=element_text(angle = 45, hjust = 1), 
        #axis.text.x=element_blank(),
        #axis.title.x=element_blank(),
        legend.position="none"
        )+
  # guides(col = guide_legend(nrows = 2)) +  # Sets legend parameters
  xlab("Date") +
  scale_x_datetime(breaks = date_breaks("2 weeks"), labels = date_format("%b %d")) +
  ylab(expression(paste("Mass. S-Meto.  ", {(g)}))) +
  scale_y_continuous(breaks = c(1, 25, 50, 100), limits = c(0, 100) )


massBal = plot_grid(massBalTop, massBalBottom, ncol = 1, nrow = 2, align = "v")


massBal_MEL <- ggplot(pgSimple) + 
  geom_line(aes(x=ti, y=value, group = variable, color=variable)) +
  
  # Themes and axes
  theme_bw() +
  theme(# axis.text.x=element_text(angle = 45, hjust = 1), 
        #axis.text.x=element_blank(),
        #axis.title.x=element_blank(),
        legend.position="top"
        )+
  labs(color = "Estimated Mass") +
  guides(col = guide_legend(ncol = 3)) +  # Sets legend parameters
  
  # xlab("Date") +
  scale_x_datetime(breaks = date_breaks("2 weeks"), labels = date_format("%b %d")) +
  ylab(expression(paste("Mass ", {(g)})))

massBal_MEL
  
```
