==> 09.0_Enrichment.Rmd <==
---
title: "Lab Enrichment"
author: "PAZ"
date: "26 septembre 2017"
output: pdf_document
---


```{r, echo=FALSE, message=FALSE, include=FALSE} 
knitr::opts_chunk$set(echo = FALSE, message=FALSE, warning = F)

Sys.setlocale("LC_ALL", "English") 

# Import lab values 
source("global.R")

library(sm)
library(vioplot)

library(dplyr)
library(tidyr)
library(zoo)
library(reshape)
library(ggplot2)
library("ggrepel")

library("plotly")
library("cowplot")
library("gridExtra")
library("Cairo")
library("GGally")
library("scales")

library("plotKML")

# Stats
library("vegan")
library("cluster")

```

## Extraction error correlation - Alteck Soils
```{r}
c = c(5, 10, 20, 35, 50, 5, 10, 20, 35, 50)
sh = c(0.993, 1.000, -0.179, 1.184, 0.810, 0.732, 0.566, 0.831, 1.008, 0.704)

cor.test(c, sh)
```
## Difference between error means - Water

Two tailed, if P < 0.05, we reject the null hypothesis that u1 = u2.

Results show that we cannot reject $H_o$ (i.e., that means are equal, thus no sig. difference exists between means and the two populations distributions do not differ.
```{r}
mQ = c(-0.118, -0.497, -0.203, -0.432, 1.062, -0.114)
mean(mQ)

envW = c(1.218, 0.966, 0.422, -0.131)
mean(c(envW, mQ))

# If P < 0.05, reject the null hypothesis that u1 = u2.
wilcox.test(mQ, envW, paired = F)

```
## Difference between error means - Soils

### Paddy vs Rouff

```{r}
paddy = c(-1.468, 0.572, -0.033, 0.378, -0.297, -0.918, 1.500, -1.047, 0.052, -0.186)
rouff = c(1.620, 1.520, 1.773, 1.534, 1.217, 0.021, 1.361, 1.170, 1.328, 1.194)

mean(paddy)
sd(paddy)
# If P < 0.05, reject the null hypothesis that u1 = u2.
wilcox.test(paddy, rouff, paired = T)
```

Result: Paddy and Rouff are significantly different.


### Paddy vs Alteck
```{r}
alteck = c(0.993, 1.000, -0.179, 1.184, 0.810, 0.732, 0.566, 0.831, 1.008, 0.704)
mean(alteck)

# If P < 0.05, reject the null hypothesis that u1 = u2.
wilcox.test(paddy, alteck, paired = T)
```

Result: Paddy and Alteck are significantly different.

### Rouff vs Alteck
```{r}
wilcox.test(rouff, alteck, paired = T)
```

Result: Rouffach and Alteck are significantly different.

### Propagated error 

Propagated error accounts for 1 SD of initial product and 1 SD from the method

```{r}
# Computed in "global.R"
propagatedError
```

## All data

```{r}
# Import raw data
enrich = read.csv2('Data/EnrichmentExp.csv', sep = ";", dec = ".", header = T) 
enrich = subset(enrich, Temp == 20)
enrich$Type = as.character(enrich$Type)

# Initial mean concentration
inital = subset(enrich, Days == 1 ) # & Type == "Biotic")
iConc = mean(inital$C.SM)

# Iniital Delta in deg experiment
iniDelta = mean(inital$Delta)
enrich$DD13 <- enrich$Delta - initialDelta
enrich$DD13m <- enrich$Delta - iniDelta

abiotic = subset(enrich, Type == "Abiotic" & Temp == 20)
biotic = subset(enrich, Type == "Biotic" & Temp == 20)

```

# Degradation rate constant ($k_{1/2}$), DT50 and DT90 (Overall)

Single frist order rate model based on:

https://www.epa.gov/pesticide-science-and-assessing-pesticide-risks/degradation-kinetics-equations

```{r}
# Get Degradation rate constant 

biotic$lnCtC0 = log(biotic$C.SM/iConc) 
deg.lm = lm(lnCtC0 ~ Days, data = biotic)
out = summary(deg.lm)

k = -as.numeric(coef(deg.lm)[2])

high = -confint(deg.lm, level = 0.95)[2,1]
low = -confint(deg.lm, level = 0.95)[2,2]
CI = high - k
se = sqrt(diag(vcov(deg.lm)))[2]

k
se[[1]]

ggplot(data = biotic, aes(y=lnCtC0, x = Days))+
  geom_point() +
  theme_minimal() +
  stat_smooth(method = "lm", formula = y ~ x, se=F)
  
DT50 = log(2)/k
DT50_high = log(2)/(k+se[[1]])
DT50_low = log(2)/(k-se[[1]])


DT90 = log(10)/k
DT90_high = log(10)/(k+se[[1]])

p_k = lmp(deg.lm)

# Overall
# DT50
DT50
DT50 - DT50_high

# DT90
DT90
DT90 - DT90_high
```

# At 20 % Moisture

```{r}
biotic_20 = subset(biotic, Temp == 20 & Theta == 20)

deg.lm = lm(lnCtC0 ~ Days, data = biotic_20)
out = summary(deg.lm)

k = -as.numeric(coef(deg.lm)[2])

high = -confint(deg.lm, level = 0.95)[2,1]
low = -confint(deg.lm, level = 0.95)[2,2]
CI = high - k
se = sqrt(diag(vcov(deg.lm)))[2]

k
se[[1]]

ggplot(data = biotic, aes(y=lnCtC0, x = Days))+
  geom_point() +
  theme_minimal() +
  stat_smooth(method = "lm", formula = y ~ x, se=F)
  
DT50 = log(2)/k
DT50_high = log(2)/(k+se[[1]])
DT50_low = log(2)/(k-se[[1]])


DT90 = log(10)/k
DT90high = log(10)/(k+se[[1]])

p_k = lmp(deg.lm)

# Overall
# DT50
DT50
DT50-DT50_high

# DT90
DT90
DT90-DT90high
```

# At 20 % Moisture

```{r}
biotic_40 = subset(biotic, Temp == 20 & Theta == 40)

deg.lm = lm(lnCtC0 ~ Days, data = biotic_40)
out = summary(deg.lm)

k = -as.numeric(coef(deg.lm)[2])

high = -confint(deg.lm, level = 0.95)[2,1]
low = -confint(deg.lm, level = 0.95)[2,2]
CI = high - k
se = sqrt(diag(vcov(deg.lm)))[2]

k
se[[1]]

ggplot(data = biotic, aes(y=lnCtC0, x = Days))+
  geom_point() +
  theme_minimal() +
  stat_smooth(method = "lm", formula = y ~ x, se=F)
  
DT50 = log(2)/k
DT50_high = log(2)/(k+se[[1]])
DT50_low = log(2)/(k-se[[1]])


DT90 = log(10)/k
DT90high = log(10)/(k+se[[1]])

p_k = lmp(deg.lm)

# Overall
# DT50
DT50
DT50-DT50_high

# DT90
DT90
DT90-DT90high
```

# Conc. vs. $\Delta \delta$ (based on ini. conc of the experimental setup)

```{r}


```


# $\Delta \delta$ vs time

```{r}



enrich$C.SMprct = (enrich$C.SM/c_ini)*100

# Test Models - DD vs Time
abiotic.lm = lm(DD13m ~ Days, data=abiotic)
summary(abiotic.lm)

biotic.lm = lm(DD13m ~ Days, data=biotic)
summary(biotic.lm)


# Function to retriece the p-value
# lmp() see file: global.r
p_abiotic = lmp(abiotic.lm)
p_biotic = lmp(biotic.lm)

cof_abiotic <- as.numeric(coef(abiotic.lm)[2])
cof_biotic <- as.numeric(coef(biotic.lm)[2])

cofa_label <- sprintf("beta == %0.3f", cof_abiotic)
pa_label <- sprintf("P == %0.2f", p_abiotic)

cofb_label <- sprintf("beta == %0.3f", cof_biotic)
pb_label <- sprintf("P < %0.4f", p_biotic)

formula = sprintf("y == beta * X")


# Draw graphs

red = "#F8766D"
blue = "#00BFC4"

exp = ggplot(data = enrich, aes(x=Days, y=DD13m, group = Type, colour = Type))+
  geom_errorbar(aes(ymin = DD13m - SD, ymax = DD13m + SD)) +
  geom_point() + 
  theme_minimal() +
  theme(legend.position = "top",
        text = element_text(size=17)) +
  stat_smooth(method = "lm", formula = y ~ x, se=F) +
  annotate("text", x = 150.0, y = 6.0, label = formula, parse = T, size = 4, colour = blue) +
  annotate("text", x = 150.0, y = 5.5, label = cofb_label, parse = T, size = 4, colour= blue) +
  annotate("text", x = 150.0, y = 5.0, label = pb_label, parse = T, size = 4, colour = blue) + 
  annotate("text", x = 150.0, y = 2.3, label = formula, parse = T, size = 4, colour = red) +
  annotate("text", x = 150.0, y = 1.8, label = cofa_label, parse = T, size = 4, colour= red) +
  annotate("text", x = 150.0, y = 1.3, label = pa_label, parse = T, size = 4, colour = red) +
  ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) 


  
exp
# Note: SD are error propagated: prop.d13C.SD
SAVE = F
PC = T
if (SAVE){
  if (PC){
    ggsave(exp, 
       filename = "D:/Documents/these_pablo/WriteUp/Images/DDvsDays.pdf", 
       device = "pdf", dpi = 600,  scale = 1, # )# ,
       width = 8.7, height = 6)
  } else {
    ggsave(exp,
     filename = "/Users/DayTightChunks/Documents/PhD/Writeups/DDvsDays.pdf", 
      device=cairo_pdf, dpi = 600,  scale = 1, # )# ,
       width = 8.7, height = 6)
  }
}
# ggsave(exp, filename = "/Users/DayTightChunks/Documents/PhD/Writeups/DegExp2.png" , width = 8.7, height = 6 )#, units = "cm", scale = 1)
```


# Delta vs Conc

```{r}
enrich = read.csv2('Data/EnrichmentExp.csv', sep = ";", dec = ".", header = T) 
enrich$Temp = as.numeric(enrich$Temp)
# enrich$Days = as.factor(enrich$Days)
abiotic = subset(enrich, Type == "Abiotic")
abiotic2020 = subset(enrich, Type == "Abiotic" & Temp == 20 & Theta == 20)
abiotic2040 = subset(enrich, Type == "Abiotic" & Temp == 20 & Theta == 40)
abiotic3020 = subset(enrich, Type == "Abiotic" & Temp == 30 & Theta == 20)
abiotic3040 = subset(enrich, Type == "Abiotic" & Temp == 30 & Theta == 40)

ini2020 = abiotic2020$C.SM[1]
ini2040 = abiotic2040$C.SM[1]
abiotic2020$C.SMprct = (abiotic2020$C.SM/ini2020)
abiotic2040$C.SMprct = (abiotic2040$C.SM/ini2040)

ini3020 = abiotic3020$C.SM[1]
ini3040 = abiotic3040$C.SM[1]
abiotic3020$C.SMprct = (abiotic3020$C.SM/ini3020)
abiotic3040$C.SMprct = (abiotic3040$C.SM/ini3040)

df1 = rbind(abiotic2020, abiotic2040)

mab <- aggregate(C.SMprct~Days, df1, mean) 
SDab = aggregate(C.SMprct~Days, df1, sd)

ab = merge(mab, SDab, by = "Days")
ab$Type = as.factor("Abiotic")

biotic2020 = subset(enrich, Type == "Biotic" & Temp == 20 & Theta == 20)
biotic2040 = subset(enrich, Type == "Biotic" & Temp == 20 & Theta == 40)
ini2020 = biotic2020$C.SM[1]
ini2040 = biotic2040$C.SM[1]
biotic2020$C.SMprct = (biotic2020$C.SM/ini2020)
biotic2040$C.SMprct = (biotic2040$C.SM/ini2040)

df2 = rbind(biotic2020, biotic2040) # , abiotic2020, abiotic2040)


mb <- aggregate(C.SMprct~Days, df2, mean) 
SDb = aggregate(C.SMprct~Days, df2, sd)

b = merge(mb, SDb, by = "Days")
b$Type = as.factor("Biotic")

aba = rbind(ab, b)

ggplot(data = subset(aba), aes(y=C.SMprct.x, x=Days, group = Type, colour = Type))+
  geom_errorbar(aes(ymin = C.SMprct.x - C.SMprct.y, ymax = C.SMprct.x + C.SMprct.y)) +
  geom_point() + 
  theme_minimal()

df3 = rbind(df2, df1)
g = ggplot(data = subset(df3), aes(y=C.SMprct, x=Days))+
  # geom_errorbar(aes(ymin = DD13 - SD, ymax = DD13 + SD)) +
  geom_point(aes(group = Type, colour = Type)) + 
  theme_minimal() +
  # stat_smooth(method = "lm", formula = y ~ x, se=F) +
  ylab(expression(paste("C"["t"]/"C"["0"], ' (-)'))) 

g

# ggsave(g, filename = "/Users/DayTightChunks/Documents/PhD/Writeups/DegExp1.png" , width = 8.7, height = 6 )#, units = "cm", scale = 1)
```


## Degradation experiments and $\varepsilon_{lab}$ derivation 

```{r}

bio = subset(enrich, Type==as.character("Biotic") & Temp == 20)


cor.test(bio$Delta , bio$C.SM)

pearson_r <- cor.test(bio$Delta , bio$C.SM)[4]
r_label <- sprintf("Pearson~r == %0.2f", pearson_r)
p_value <- cor.test(bio$Delta, bio$C.SM)[3]

if (p_value < 0.0001){
  p_label <- "(P < 0.001)"
} else if (p_value < 0.001) {
  p_label <- "(P < 0.001)"
} else if (p_value < 0.015) {
  p_label <- ("P < 0.01")
} else {
  p_label <- "Check significance"
}

b = ggplot(data = bio, aes(x=C.SM, y=DD13))+
  geom_errorbar(aes(ymin = DD13 - SD, ymax = DD13 + SD)) +
  geom_point(aes(group = Theta, size = Days)) +
  
  # geom_errorbarh(aes(xmin = Conc.mug.g.dry.soil - Conc.ComSoil.SD, xmax = Conc.mug.g.dry.soil + Conc.ComSoil.SD)) +
  stat_smooth(data = bio, 
              aes(x=C.SM, y=DD13), method = "lm", formula = y ~ x, se=F) +
  scale_size_continuous(range = c(1, 5), breaks= c(1, 10, 50, 100, 200), limits = c(1, 200)) + 
  scale_y_continuous(breaks=c(0, 1, 2, 3 , 4 ,5, 6, 7 , 8) ) + 
  # geom_point(aes(group = ID, size = timeSinceApp.NoSo)) + # , colour = Source)) +  # , shape = ngC.Label)) +
  theme_minimal() +
  theme(legend.position = "top" ,
        text = element_text(size=23)
        ) +
  labs(size="   Days after spiking") + #, shape = "Mass Carbon") +
  ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  xlab(expression(paste("S-met Soil Concentration  ", {({mu}*g / g~dry~wt.)}))) +
  annotate("text", x = 2.0, y = 4.7, label = as.character(r_label), 
           size = 5,
           parse = T) + 
  annotate("text", x = 2.0, y = 4.0, label = p_label, parse = T, size = 5) +
  annotate("rect", xmin=0, xmax=3.5, ymin=-0.9, ymax=0.9, alpha=0.2)
```

# Rayleigh (20 \textdegree C, $\theta$: 20 \& 40)

```{r}
bio$DD13 <- bio$Delta - delta_ini
bio$yRaleigh <- log((1000+delta_ini+bio$DD13)/(1000+delta_ini))
bio$xRaleigh <- log(bio$C.SM/c_ini)
expModel<-lm(yRaleigh~xRaleigh, data= bio) 

cofsoil <- as.numeric(coef(expModel)[2]*1000)
minX <- confint(expModel, "xRaleigh", level = 0.95)[1]*1000
maxX <- confint(expModel, "xRaleigh", level = 0.95)[2]*1000
se <- summary(expModel)$coef[[4]]*1000

e_label <- sprintf("epsilon == %0.3f", cofsoil)
cond_label = paste("theta:~20~+~40~and~degree~C:20")

CI95 = maxX - cofsoil

ggplot(data = subset(bio,  !is.na(yRaleigh) ), aes(x=xRaleigh, y=yRaleigh)) +
  geom_point() +
  stat_smooth(method = "lm", formula = y ~ x, se=F) +
  annotate("text", x = -1.5, y = 0.007, label = as.character(e_label), parse = T, size = 3.5) +
  annotate("text", x = -1.5, y = 0.006, label = cond_label, parse = T, size = 3.5) +
  theme_bw()

summary(expModel)
```

# Rayleigh (20 \textdegree C, $\theta$: 20)

```{r}
bio = subset(enrich, Type==as.character("Biotic") & Temp == 20 & Theta == 20)
bio$DD13 <- bio$Delta - delta_ini
bio$yRaleigh <- log((1000+delta_ini+bio$DD13)/(1000+delta_ini))
bio$xRaleigh <- log(bio$C.SM/c_ini)
expModel<-lm(yRaleigh~xRaleigh, data= bio) 

cofsoil <- as.numeric(coef(expModel)[2]*1000)
minX <- confint(expModel, "xRaleigh", level = 0.95)[1]*1000
maxX <- confint(expModel, "xRaleigh", level = 0.95)[2]*1000
se <- summary(expModel)$coef[[4]]*1000

e_label <- sprintf("epsilon == %0.3f", cofsoil)
cond_label = paste("theta:~20~and~degree~C:20")

CI95 = maxX - cofsoil

ggplot(data = subset(bio,  !is.na(yRaleigh) ), aes(x=xRaleigh, y=yRaleigh)) +
  geom_point() +
  stat_smooth(method = "lm", formula = y ~ x, se=F) +
  annotate("text", x = -1.5, y = 0.007, label = as.character(e_label), parse = T, size = 3.5) +
  annotate("text", x = -1.5, y = 0.006, label = cond_label, parse = T, size = 3.5) +
  theme_bw()

summary(expModel)
```

# Rayleigh (20 \textdegree C, $\theta$: 40)

```{r}
bio = subset(enrich, Type==as.character("Biotic") & Temp == 20 & Theta == 40)

bio$DD13 <- bio$Delta - delta_ini
bio$yRaleigh <- log((1000+delta_ini+bio$DD13)/(1000+delta_ini))
bio$xRaleigh <- log(bio$C.SM/c_ini)
expModel<-lm(yRaleigh~xRaleigh, data= bio) 

cofsoil <- as.numeric(coef(expModel)[2]*1000)
minX <- confint(expModel, "xRaleigh", level = 0.95)[1]*1000
maxX <- confint(expModel, "xRaleigh", level = 0.95)[2]*1000
se <- summary(expModel)$coef[[4]]*1000

e_label <- sprintf("epsilon == %0.3f", cofsoil)
cond_label = paste("theta:~40~and~degree~C:20")

CI95 = maxX - cofsoil

ggplot(data = subset(bio,  !is.na(yRaleigh) ), aes(x=xRaleigh, y=yRaleigh)) +
  geom_point() +
  stat_smooth(method = "lm", formula = y ~ x, se=F) +
  annotate("text", x = -.5, y = 0.005, label = as.character(e_label), parse = T, size = 3.5) +
  annotate("text", x = -.5, y = 0.004, label = cond_label, parse = T, size = 3.5) +
  theme_bw()

summary(expModel)
```



## Abiotic data

### Abiotic - Rayleigh (20 \textdegree C, $\theta$: 20 \& 40)

```{r}
abiotic = subset(enrich, Type==as.character("Abiotic") & Temp == 20)

abiotic$DD13 <- abiotic$Delta - delta_ini
pearson_r <- cor.test(abiotic$Delta , abiotic$C.SM)[4]
r_label <- sprintf("Pearson~r == %0.2f", pearson_r)
p_value <- cor.test(abiotic$Delta, abiotic$C.SM)[3]

if (p_value < 0.0001){
  p_label <- "(P < 0.001)"
} else if (p_value < 0.001) {
  p_label <- "(P < 0.001)"
} else if (p_value < 0.015) {
  p_label <- ("P < 0.01")
} else {
  p_label <-  sprintf("P == %0.2f", p_value)
}

ab = ggplot(data = abiotic, aes(x=C.SM, y=DD13))+
  geom_errorbar(aes(ymin = DD13 - SD, ymax = DD13 + SD)) +
  geom_point(aes(group = Theta, size = Days)) +
  
  # geom_errorbarh(aes(xmin = Conc.mug.g.dry.soil - Conc.ComSoil.SD, xmax = Conc.mug.g.dry.soil + Conc.ComSoil.SD)) +
  stat_smooth(data = abiotic, 
              aes(x=C.SM, y=DD13), method = "lm", formula = y ~ x, se=F) +
  scale_size_continuous(range = c(1, 5), breaks= c(1, 10, 50, 100, 200), limits = c(1, 200)) + 
  scale_y_continuous(breaks=c(0, 1, 2, 3 , 4 ,5, 6, 7 , 8) ,  limits = c(-0.9, 8)) + 
  # geom_point(aes(group = ID, size = timeSinceApp.NoSo)) + # , colour = Source)) +  # , shape = ngC.Label)) +
  theme_minimal() +
  theme(legend.position = "top" ,
        text = element_text(size=23)
        ) +
  labs(size="   Days after spiking") + #, shape = "Mass Carbon") +
  ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  xlab(expression(paste("S-met Soil Concentration  ", {({mu}*g / g~dry~wt.)}))) +
  annotate("text", x = 3.0, y = 4.7, label = as.character(r_label), 
           size = 5,
           parse = T) + 
  annotate("text", x = 3.0, y = 4.0, label = p_label, parse = T, size = 5) +
  annotate("rect", xmin=0, xmax=max(abiotic$C.SM), ymin=-0.9, ymax=0.9, alpha=0.2)
```


```{r}
# Import raw data
abiotic$yRaleigh <- log((1000+delta_ini+abiotic$DD13)/(1000+delta_ini))
abiotic$xRaleigh <- log(abiotic$C.SM/c_ini)
expModel<-lm(yRaleigh~xRaleigh, data= abiotic) 


cofsoil <- as.numeric(coef(expModel)[2]*1000)
minX <- confint(expModel, "xRaleigh", level = 0.95)[1]*1000
maxX <- confint(expModel, "xRaleigh", level = 0.95)[2]*1000
se <- summary(expModel)$coef[[4]]*1000

e_label <- sprintf("epsilon == %0.3f", cofsoil)
cond_label = paste("theta:~20~+~40~and~degree~C:20")

CI95 = maxX - cofsoil

ggplot(data = subset(abiotic,  !is.na(yRaleigh) ), aes(x=xRaleigh, y=yRaleigh)) +
  geom_point() +
  stat_smooth(method = "lm", formula = y ~ x, se=F) +
  #annotate("text", x = -1.5, y = 0.007, label = as.character(e_label), parse = T, size = 3.5) +
  #annotate("text", x = -1.5, y = 0.006, label = cond_label, parse = T, size = 3.5) +
  theme_bw()

summary(expModel)

```


### Abiotic - Rayleigh (20 \textdegree C, $\theta$: 20)

```{r}

abiotic = subset(enrich, Type==as.character("Abiotic") & Temp == 20 & Theta == 20)

abiotic$DD13 <- abiotic$Delta - delta_ini
abiotic$yRaleigh <- log((1000+delta_ini+abiotic$DD13)/(1000+delta_ini))
abiotic$xRaleigh <- log(abiotic$C.SM/c_ini)
expModel<-lm(yRaleigh~xRaleigh, data= abiotic) 


cofsoil <- as.numeric(coef(expModel)[2]*1000)
minX <- confint(expModel, "xRaleigh", level = 0.95)[1]*1000
maxX <- confint(expModel, "xRaleigh", level = 0.95)[2]*1000
se <- summary(expModel)$coef[[4]]*1000

e_label <- sprintf("epsilon == %0.3f", cofsoil)
cond_label = paste("theta:~20~+~40~and~degree~C:20")

CI95 = maxX - cofsoil

ggplot(data = subset(abiotic,  !is.na(yRaleigh) ), aes(x=xRaleigh, y=yRaleigh)) +
  geom_point() +
  stat_smooth(method = "lm", formula = y ~ x, se=F) +
  #annotate("text", x = -1.5, y = 0.007, label = as.character(e_label), parse = T, size = 3.5) +
  #annotate("text", x = -1.5, y = 0.006, label = cond_label, parse = T, size = 3.5) +
  theme_bw()

summary(expModel)

```

### Abiotic - Rayleigh (20 \textdegree C, $\theta$: 40)

```{r}

abiotic = subset(enrich, Type==as.character("Abiotic") & Temp == 20 & Theta == 40)

abiotic$DD13 <- abiotic$Delta - delta_ini
abiotic$yRaleigh <- log((1000+delta_ini+abiotic$DD13)/(1000+delta_ini))
abiotic$xRaleigh <- log(abiotic$C.SM/c_ini)
expModel<-lm(yRaleigh~xRaleigh, data= abiotic) 


cofsoil <- as.numeric(coef(expModel)[2]*1000)
minX <- confint(expModel, "xRaleigh", level = 0.95)[1]*1000
maxX <- confint(expModel, "xRaleigh", level = 0.95)[2]*1000
se <- summary(expModel)$coef[[4]]*1000

e_label <- sprintf("epsilon == %0.3f", cofsoil)
cond_label = paste("theta:~20~+~40~and~degree~C:20")

CI95 = maxX - cofsoil

ggplot(data = subset(abiotic,  !is.na(yRaleigh) ), aes(x=xRaleigh, y=yRaleigh)) +
  geom_point() +
  stat_smooth(method = "lm", formula = y ~ x, se=F) +
  #annotate("text", x = -1.5, y = 0.007, label = as.character(e_label), parse = T, size = 3.5) +
  #annotate("text", x = -1.5, y = 0.006, label = cond_label, parse = T, size = 3.5) +
  theme_bw()

summary(expModel)

```
