==> 09.2_LutzModel.Rmd <==
---
title: "Lutz Model Comparison"
author: "PAZ"
date: "06/04/2017"
output: pdf_document
bibliography: library.bib
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning=FALSE, message = FALSE)
Sys.setlocale("LC_ALL", "English")
```

# Introduction

This notebook compares Lutz et al [@Lutz2013] simulations to catchment outlet data. 

# Files

- **WaterDay_R.csv** (Book 02)
- **Data/OutletData4Lutz_R.csv**  (Book 9.1)
- **PCA4Lutz_R.csv** (Book 10.3)

# Packages

```{r, warning=FALSE,}
library(sm)
library(vioplot)

library(dplyr)
library(tidyr)
library(zoo)
library(reshape)
library(ggplot2)
library("ggrepel")

library("plotly")
library("cowplot")
library("gridExtra")
library("Cairo")
library("GGally")
library("scales")

library("plotKML")

# Stats
library("vegan")
library("cluster")

# Saving a xlxs file
# library(xlsx)
```

# Lutz summary

## Steady state

- Initial constant concentration (not single application) and recharge input boundaries were maintained until concentrations at the outlet reached a steady state.
- Recharge rate of 250 mm/yr

### Steady state results

- Mean travel time for groundwater was 6.7 yrs (5 yr for the pesticide)
- Isotopes are enriched with distance from the source area. Max ($\Delta\delta$ $= 4 ‰$) is reached at the hillslope end in the shallow areas close to outlet. 


### Relationships to field data

- Extent of isotope shifts are comparable to the staedy state simulation but only near the end of the season in June. Early season signatures are closer to initial product signatures, thus the role of preferential flow pathways is likely to be important during the early season. 

## Extreme rainfall

- A single event to characterize how the hillslope reacts when an extreme event leads to surface run-off. 
- Rainfall was applied at a uniform intensity of 60 mm/h for 30 min.
- Recharge rate was assumed as during steady state.
- Concentration at the source area wa kept constant (i.e. not single pesticide application)

### Extreme rainfall comparison

- Compares well with late season, when crusting has been well developed


## Transient state

- Incorporates periods of baseflow and extreme rainfall events to evaluate how concentrations and isotopes respond throughout the year.
- Use of daily rainfall and ETP values for a metereological station for 1 yr.
- Data includes one major event with a maximum intensity of 27 mm/h, total amount of 106.4 mm and lasting 20.8 hrs.
- Same data was applied to 20 consecutive simulation years (i.e. to faciliatate hydrological interpretation)
- Pesticide was applied once a year on a dry day in spring (day 100; 11 April); 

### Transient state comparison

- Rainfall events reduce concentrations at the outlet due to dilution (not clear if runoff is considered in this scenario, but don't think so); 
- and continue to influence concentrations at the outlet because they dilute also the sub-surface
- So concentration increases are associated to dry periods, where baseflow contributes more to the outlet discharge, without fresh water dilution
- Pesticide applied during the dry season produces no sudden decrease in isotope signatures at the outelt but rather a slow but steady shift in outlet isotope signatures in the direction of the applied product. Degradation at the outlet is highest durign this period, reflecting discharge of groundwater or baseflow components wwith longer reactivity periods. 


### Comparison to Transient

- Our dry period did not lead to concentration increases.
- Concentration increases where rather seen as a result of large rainfall events



## Data preparation for comparison to Lutz et al. (2013)

To choose field events, cummulative rainfalls over a 30 minute interval exceeding at least 4 mm/hr were selected. The start of the event is considered to be the most recent maximum rainfall intensity associated to the aliquots making up the composite sample. Therefore, signatures represent the flow proportional average of the discharge taking place at the outlet over the course of the event.     

```{r}

wstidier <- read.csv2("Data/OutletData4Lutz_R.csv", header = TRUE)
                      # na.strings=c('#DIV/0!', '', 'NA'), header = TRUE)
wstidier$Date <- as.character(wstidier$Date)
wstidier$Date <- as.POSIXct(strptime(wstidier$Date, 
                                          "%Y-%m-%d %H:%M", tz="EST")) 
names(wstidier)
wstidier = subset(wstidier, Source == "Outlet")
wstidier = wstidier[, c( "Date", "Event", "Events", "measure", "SD")]

# Import the PCA data used to cluster events (might be better than above)
yClust <- read.csv2("Data/PCA4Lutz_R.csv", header = TRUE)
names(yClust)[names(yClust) == "Date.ti.x"] <- "Date"
yClust$Date.ti.y <- NULL
yClust$Date <- as.POSIXct(strptime(yClust$Date, 
                                          "%Y-%m-%d %H:%M", tz="EST")) 
names(yClust)

# 30 minute rainfall
rainDay30min <- read.csv2("Data/30minRain.csv", header = T, dec =".")
rainDay30min$Time <- as.character(rainDay30min$Time) 
rainDay30min$Date <-  as.POSIXct(strptime(rainDay30min$Time, 
                                          "%d/%m/%Y %H:%M", tz="EST")) 

rainDay30min$Time <- NULL
rainDay30min$Cumm.mm <- as.numeric(rainDay30min$Cumm.mm)

# Convert to mm/hr
rainDay30min <- subset(rainDay30min, Cumm.mm > 0)
rainDay30min$mm.hr <- rainDay30min$Cumm.mm*(60/30) # mm/min -> mm/hr
rainDay30min$Cumm.mm <- NULL

# Sampler data
sampler <- read.csv2("Data/SamplerFormatted_R.csv", header = T)
sampler$Date <- as.POSIXct(strptime(sampler$Date, 
                                          "%Y-%m-%d %H:%M", tz="EST")) 

rainSample <- merge(rainDay30min, sampler, by = "Date", all = T)
# anyDuplicated(rainSample$Date)

yrs <- merge(yClust, rainSample, by = "Date", all = T )
yrs$Events <- na.locf(yrs$Events) 
yrs$mm.hr[1] <- 0
yrs$mm.hr = ifelse(is.na(yrs$mm.hr), 0, yrs$mm.hr)
yrs$Cluster = na.locf(yrs$Cluster)

yrs0 = yrs %>%
  group_by(Events) %>%
  dplyr::summarize(Cluster = max(Cluster),
                   max.mm = max(mm.hr),
                   Label = first(Label),
                   EventLabel = first(EventLabel),
                   tmax = Date[which.max(mm.hr)],
                   ti = first(Date), # 
                   tf = last(Date),
                   DD.diss = first(DD13C.diss),
                   DD.nrm = first(DD.diss.nrm),
                   DD.nrmSD = first(DD.diss.nrmSD)) 

yrs1 = merge(yrs0, wstidier, by.x = "ti", by.y = "Date", all = T)

compare <- subset(yrs1, Cluster != 1 & !is.na(measure) ) 
compare$Duration.Hrs = 
  as.numeric(difftime(compare$tf, compare$ti, units = "hours"), units = "hours")

compare$DtoMax = 
  as.numeric(difftime(compare$tmax, compare$ti, units = "hours"), units = "hours")

compare$DfromMax = 
  as.numeric(difftime(compare$tf, compare$tmax, units = "hours"), units = "hours")

compare$Approach = "Outlet, 2016"

evA = subset(compare, Label == "A")
evA$Hrs = cumsum(evA$Duration.Hrs)
evA$Hrs[1] = min(evA$DfromMax[1], evA$DtoMax[1])

evB = subset(compare, Label == "B")
evB$Hrs = cumsum(evB$Duration.Hrs)
evB$Hrs[1] = min(evB$DfromMax[1], evB$DtoMax[1])

evC = subset(compare, Label == "C")
evC$Hrs = cumsum(evC$Duration.Hrs)
evC$Hrs[1] = min(evC$DfromMax[1], evC$DtoMax[1])

evD = subset(compare, Label == "D")
evD$Hrs = cumsum(evD$Duration.Hrs)
evD$Hrs[1] = min(evD$DfromMax[1], evD$DtoMax[1])

evE = subset(compare, Label == "E")
evE$Hrs = cumsum(evE$Duration.Hrs)
evE$Hrs[1] = min(evE$DfromMax[1], evE$DtoMax[1])

evF = subset(compare, Label == "F")
evF$Hrs = cumsum(evF$Duration.Hrs)
evF$Hrs[1] = min(evF$DfromMax[1], evF$DtoMax[1])

evG = subset(compare, Label == "G")
evG$Hrs = cumsum(evG$Duration.Hrs)
evG$Hrs[1] = min(evG$DfromMax[1], evG$DtoMax[1])

events <- rbind(evA, evB, evC, evD, evE, evF, evG)

events$HasLabel = ifelse(events$Label == "A", "yes",
                         ifelse(events$Label == "B", "yes",
                                ifelse(events$Label == "E", "yes", "no")
                                )
                         )

events$Label = as.character(events$Label)
events$Label =ifelse(events$Label == "A", "A (May 12)", 
                     ifelse(events$Label == "B", "B (May 29)*",
                            ifelse(events$Label == "C", "C (June 2)",
                                   ifelse(events$Label == "D", "D (June 3)",
                                          ifelse(events$Label == "E", "E (June 4)",
                                                 ifelse(events$Label == "F", "F (June 14)",
                                                        ifelse(events$Label == "G", "G (June 25)", events$Label)
                                                        )
                                                 )
                                          )
                                   )
                            )
                     )
                            

events <- events[c("Hrs", "max.mm", "measure", "SD" , "DD.nrm", "DD.nrmSD",
                   "EventLabel", "Cluster",
                   "Label", "Approach", "HasLabel")]
names(events) = c("Duration.hr", "mm.hr", "measure", "SD" ,  "DD.nrm", "SD.nrm",
                  "EventLabel", "Cluster",
                  "Event", "Approach", "HasLabel")
events$Event  = as.factor(events$Event )
events$Cluster = as.factor(events$Cluster)
names(events)

# Lutz reproduction
Duration.hr <- c(0, 
                 24*.1, 
                 24*.2, 
                 24*.3, 
                 24*.4, 
                 24*.5, 
                 15, 
                 24, 30)

iniLutz = -30
measure <- c(-30 - iniLutz, 
             -28.7 - iniLutz, 
             -27.5 - iniLutz, 
             -27 - iniLutz, 
             -26.9 - iniLutz, 
             -26.9 - iniLutz,
             -26.9 - iniLutz,
             -26.9 - iniLutz,
             -26.9 - iniLutz)

DD.nrm = c((-30 - iniLutz)/(-26.9 - iniLutz), 
           (-28.7 - iniLutz)/(-26.9 - iniLutz),
           (-27.5 - iniLutz)/(-26.9 - iniLutz),
           (-27 - iniLutz)/(-26.9 - iniLutz),
           (-26.9 - iniLutz)/(-26.9 - iniLutz), 
           (-26.9 - iniLutz)/(-26.9 - iniLutz),
           (-26.9 - iniLutz)/(-26.9 - iniLutz),
           (-26.9 - iniLutz)/(-26.9 - iniLutz),
           (-26.9 - iniLutz)/(-26.9 - iniLutz)
           ) 
             
            

reps = length(measure)
mm.hr <- c(60, rep(0, reps - 1))
SD <- rep(NA, reps)
SD.nrm <- rep(NA, reps)
EventLabel <- rep(NA, reps)
Event =  rep("Model hillslope", reps)
Cluster =  NA

HasLabel = rep(NA, reps)
Approach <- rep("Model hillslope", reps)

eventLutz <- data.frame(Duration.hr, mm.hr, measure, SD, DD.nrm, SD.nrm,
                        EventLabel, Cluster,
                        Event, Approach, HasLabel)
eventLutz$Cluster <- as.factor(eventLutz$Cluster)

allEvents <- rbind(eventLutz, events)
# allEvents$DD.nrm = ifelse(allEvents$DD.nrm == 0, allEvents$measure, allEvents$DD.nrm)


LutzEvents <- ggplot(data = allEvents, 
                     aes(x= Duration.hr,
                         y=DD.nrm) )+
                         # y=measure) )+
  # theme_bw() +
  theme_minimal() +
  # scale_y_continuous(limits = c(0, 1.14)) +
  theme(text = element_text(size=17) #,
        #legend.position = 'top'
        ) +
  #guides(col = guide_legend(nrow =  3),
  #       title.position = "top") + 
  geom_errorbar(data = allEvents, 
                aes(ymin = DD.nrm-SD.nrm, ymax = DD.nrm+SD.nrm, 
                # aes(ymin = measure-SD, ymax = measure+SD, 
                    colour = Event),
                 # width=.2 , # ) + #,                    # Width of the error bars
                #position=position_dodge(.5), 
                size=0.2) +
  geom_line(data=subset(allEvents, Event == "Model hillslope" ), aes(colour = Event), linetype = "dashed") +
  geom_point(aes(size = mm.hr, colour = Event) #, 
                 # shape = Cluster)
             ) +
  # ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  ylab(expression(paste({Delta~delta}^"13","C" ['out'], ' / ', {Delta~delta}^"13" ,"C" ['bulk']))) +
  xlab("Time after max. rainfall (hrs.)") +
  scale_size_continuous(range = c(1, 6), breaks= c(0, 5, 20, 60), limits = c(0, 60)) +
  guides(size = guide_legend(order = 4, 
                             title=expression("Rainfall " ~ (mm/h) ), 
                             ncol=1, title.position = "top", title.vjust = .26
                              )) +
  geom_text_repel(data=subset(allEvents, HasLabel == "yes"), 
                  # aes(x= Duration.hr, y=measure,  label= EventLabel),
                  aes(x= Duration.hr, y=DD.nrm,  label= EventLabel),
                 arrow = arrow(length = unit(0.005, 'npc'), type = "closed"),
                 force = 1, 
                 point.padding = unit(0.8, 'lines'), 
                 max.iter = 2e3,
                 nudge_x = 1,
                 show.legend = FALSE)#+
  #stat_smooth(data=subset(allEvents, Event != "Simulation"), aes(x= Duration.hr, y=measure, colour = Approach), 
  #            method = "lm", formula = y ~ poly(x, 2), se = T, alpha = 0.2, size=0.2) #+
  #theme(axis.title.x = element_blank() ) +
  # facet_wrap(~ Approach)#, scale="free") 

LutzEvents
  
#
# ggsave(LutzEvents, filename = "images/LutzEventsNRM.png", width = 8, height = 5, units = "in", scale = 1)

# PC
ggsave(LutzEvents, 
      filename = "D:/Documents/these_pablo/WriteUp/Alteck_PNAS_LaTeX/images/lutzNRM.pdf", 
      device = cairo_pdf, dpi = 600,  scale = 1, # )# ,
       width = 7, height = 6)

# MAC
#  ggsave(LutzEvents, 
#    filename = "/Users/DayTightChunks/Documents/PhD/Writeups/PNAS/Alteck_PNAS_LaTeX/images/lutzNRM.pdf", 
#      device = cairo_pdf, dpi = 600,  scale = 1, # )# ,
#       width = 7, height = 6)
```

## Alternative Lutz Graph

```{r}
LutzEvents2 <- ggplot(data = allEvents, aes(x= Duration.hr, y=measure) )+
  # theme_bw() +
  theme_minimal() +
  theme(text = element_text(size=17)) + 
  geom_errorbar(aes(ymin = measure-SD, ymax = measure+SD,  colour = Cluster),
                 # width=.2 , # ) + #,                    # Width of the error bars
                #position=position_dodge(.5), 
                size=0.2) +
  geom_line(data=subset(allEvents, Event == "Simulation" ), aes(colour = Cluster), linetype = "dashed") + 
  geom_point(aes(size = mm.hr, colour = Cluster)) +
  ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  xlab("Time after max. rainfall (hrs.)") +
  scale_size_continuous(range = c(1, 5), breaks= c(0, 5, 20, 60), limits = c(0, 60)) +
  guides(size = guide_legend(order = 4, 
                             title=expression("Intensity " ~ (mm/h) ), 
                             ncol=1, title.position = "top", title.vjust = .26
                              )) +
  geom_text_repel(data=subset(allEvents, HasLabel == "yes"), aes(x= Duration.hr, y=measure,  label= EventLabel),
                 arrow = arrow(length = unit(0.005, 'npc'), type = "closed"),
                 force = 1, 
                 point.padding = unit(0.8, 'lines'), 
                 max.iter = 2e3,
                 nudge_x = 1,
                 show.legend = FALSE)#+
  #stat_smooth(data=subset(allEvents, Event != "Simulation"), aes(x= Duration.hr, y=measure, colour = Approach), 
  #            method = "lm", formula = y ~ poly(x, 2), se = T, alpha = 0.2, size=0.2) #+
  #theme(axis.title.x = element_blank() ) +
  # facet_wrap(~ Approach)#, scale="free") 

# ggsave(LutzEvents2, filename = "images/LutzEvents2.png", width = 8, height = 5, units = "in", scale = 1)
```
