==> 09.3_BarPlots.Rmd <==
---
title: "Bar Charts"
author: "PAZ"
date: "6 juillet 2017"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning=FALSE, message = FALSE)
Sys.setlocale("LC_ALL", "English")
```

# Introduction

Degradation extent is calculated based on closed and open system Rayleigh equations

# Packages

```{r, warning=FALSE,}
library(sm)
library(vioplot)

library(dplyr)
library(tidyr)
library(zoo)
library(reshape)
library(ggplot2)
library("ggrepel")

library("plotly")
library("cowplot")
library("gridExtra")
library("Cairo")
library("GGally")
library("scales")

library("plotKML")

# Stats
library("vegan")
library("cluster")

# Saving a xlxs file
# library(xlsx)


# Saving images?
PC = F
SAVE = F
```


# Lab parameters

```{r}
source("global.R")
```

# Import Soils & Waters

```{r}
sw <- read.csv2("Data/MassBalance_R.csv")
sw$ti <- as.POSIXct(strptime(sw$ti, "%Y-%m-%d %H:%M", tz="EST"))

colnames(sw)[colnames(sw) == "ti"] <- "Date.ti"

names(sw)
keepMB <- c("Date.ti", "ID.N", "Duration.Hrs",
            # Theoretical
            "Appl.Mass.g",  "Appl.Mass.g.OT",
            "CumAppMass.g",  
            "CumAppMass.g.OT",
            "CumAppMass.g.N.OT", "CumAppMass.g.T.OT", "CumAppMass.g.S.OT",
            # MB soils 
            "CatchMassSoil.g", "CatchMassSoil.g.SD",
            "BulkCatch.Conc", 
            # MB Waters
            "CumOutSmeto.g", "CumOutMELsm.g",
            "TotSMout.g", "TotSMout.g.SD" , 
            "MELsm.g", "MELsm.g.SD",
            # CSIA Soils
            "BulkCatch.d13" , "BulkCatch.d13.SD", 
            # For maximum soil degradation# "DD13C.South" , 
            # CSIA Waters
            "diss.d13C" , "SD.d13C"
            )

sw <- sw[, (names(sw) %in% keepMB)]
```

## Theoretical calculations

```{r}
# Half-life calculations (days)
# See global.r for DT50's

# Track each of the main applications (4 major applications considered)
# Caclulate degradation due to cumulative days based on each individual application
sw$No_First <- ifelse(sw$Appl.Mass.g.OT == 0, NA, sw$Appl.Mass.g.OT)
sw$No_Second <- sw$No_First
sw$No_Second[1] <- 0
sw$No_Third <- sw$No_Second
sw$No_Fourth <- sw$No_Second

# Delete the 3rd and 4th observation in "Second" 
sw$No_Second[which(!is.na(sw$No_Second))[4]] <- NA
sw$No_Second[which(!is.na(sw$No_Second))[3]] <- NA
sw$No_Second <- na.locf(sw$No_Second) # fill the cumulative for the second only

sw$No_Third[which(!is.na(sw$No_Third))[4]] <- NA
sw$No_Third[which(!is.na(sw$No_Third))[2]] <- NA
sw$No_Third <- na.locf(sw$No_Third)

sw$No_Fourth[which(!is.na(sw$No_Fourth))[3]] <- NA
sw$No_Fourth[which(!is.na(sw$No_Fourth))[2]] <- NA
sw$No_Fourth <- na.locf(sw$No_Fourth)

sw$No_First[which(!is.na(sw$No_First))[4]] <- NA
sw$No_First[which(!is.na(sw$No_First))[3]] <- NA
sw$No_First[which(!is.na(sw$No_First))[2]] <- NA
sw$No_First <- na.locf( sw$No_First )

# Compute cumulative time for first, second, third and fourth applications
sw$CumDays_First <- cumsum(sw$Duration)/24

sw$dt_Second <- ifelse(sw$No_Second == 0, 0, sw$Duration)
sw$CumDays_Second <- cumsum(sw$dt_Second)/24
sw$dt_Second <- NULL

sw$dt_Third <- ifelse(sw$No_Third == 0, 0 , sw$Duration)
sw$CumDays_Third <- cumsum(sw$dt_Third)/24
sw$dt_Third <- NULL

sw$dt_Fourth <- ifelse(sw$No_Fourth == 0, 0 , sw$Duration)
sw$CumDays_Fourth <- cumsum(sw$dt_Fourth)/24
sw$dt_Fourth <- NULL

# Based on half-life = 29d
sw$remain_1st_29d <- sw$No_First*(0.5)^(sw$CumDays_First/median_half)
sw$remain_2nd_29d <- sw$No_Second*(0.5)^(sw$CumDays_Second/median_half)
sw$remain_3rd_29d <- sw$No_Third*(0.5)^(sw$CumDays_Third/median_half)
sw$remain_4th_29d <- sw$No_Fourth*(0.5)^(sw$CumDays_Fourth/median_half)

# Based on half-life = 46d
sw$remain_1st_46d <- sw$No_First*(0.5)^(sw$CumDays_First/min_half)
sw$remain_2nd_46d <- sw$No_Second*(0.5)^(sw$CumDays_Second/min_half)
sw$remain_3rd_46d <- sw$No_Third*(0.5)^(sw$CumDays_Third/min_half)
sw$remain_4th_46d <- sw$No_Fourth*(0.5)^(sw$CumDays_Fourth/min_half)

# Based on half-life = 12d
sw$remain_1st_12d <- sw$No_First*(0.5)^(sw$CumDays_First/max_half)
sw$remain_2nd_12d <- sw$No_Second*(0.5)^(sw$CumDays_Second/max_half)
sw$remain_3rd_12d <- sw$No_Third*(0.5)^(sw$CumDays_Third/max_half)
sw$remain_4th_12d <- sw$No_Fourth*(0.5)^(sw$CumDays_Fourth/max_half)

sw$remainTheo.measure <- ((sw$remain_1st_29d + sw$remain_2nd_29d +
                                sw$remain_3rd_29d + sw$remain_4th_29d)/sw$CumAppMass.g)*100
sw$remainTheo.SD1 <- ((sw$remain_1st_46d + sw$remain_2nd_46d +
                                sw$remain_3rd_46d + sw$remain_4th_46d)/sw$CumAppMass.g)*100
sw$remainTheo.SD2 <- ((sw$remain_1st_12d + sw$remain_2nd_12d +
                                sw$remain_3rd_12d + sw$remain_4th_12d)/sw$CumAppMass.g)*100

# Assumed 1 day of photolysis, half-life = 8d 
# k = 0.087 1/days
# t_{1/2} = ln(2)/k = 8 days
sw$remainTheo.photo <- ifelse(sw$Appl.Mass.g > 0, sw$Appl.Mass.g*(0.5)^(1/8), sw$Appl.Mass.g)
sw$remainTheo.photo <- cumsum(sw$remainTheo.photo)
sw$remainTheo.photo.prct <- (sw$remainTheo.photo/sw$CumAppMass.g)*100
sw$degTheo.photo <- 100 - sw$remainTheo.photo.prct

sw$degTheo.measure <- 100 - sw$remainTheo.measure
sw$degTheo.SD1 <- 100 - sw$remainTheo.SD2
sw$degTheo.SD2 <- 100 - sw$remainTheo.SD1

colnames(sw)

theo = sw[,c("Date.ti", "CumAppMass.g.OT", 
             "remain_1st_29d" , "remain_2nd_29d" , "remain_3rd_29d", "remain_4th_29d",
             "remain_1st_46d", "remain_2nd_46d", "remain_3rd_46d", "remain_4th_46d", 
             "remain_1st_12d", "remain_2nd_12d", "remain_3rd_12d", "remain_4th_12d"
             )]
theo$rem29d = (theo$remain_1st_29d + theo$remain_2nd_29d + theo$remain_3rd_29d + theo$remain_4th_29d)/theo$CumAppMass.g.OT
theo$rem46d = (theo$remain_1st_46d + theo$remain_2nd_46d + theo$remain_3rd_46d + theo$remain_4th_46d)/theo$CumAppMass.g.OT
theo$rem12d = (theo$remain_1st_12d + theo$remain_2nd_12d + theo$remain_3rd_12d + theo$remain_4th_12d)/theo$CumAppMass.g.OT

theo = theo[, c("Date.ti", "rem12d", "rem29d", "rem46d")]

mtheo = melt(theo, id=c("Date.ti"))
ggplot(data = mtheo, aes(x=Date.ti, y=value, colour=variable))+
  geom_line() +
  theme_minimal()

sfo = theo
sfo$Days = seq.int(nrow(sfo))
ini = sw$CumAppMass.g.OT[nrow(sw)-1]
sfo$Mtd21 = ini*exp((-log(2)/median_half)*(sfo$Days))
sfo$Mtd8 = ini*exp((-log(2)/max_half)*(sfo$Days))
sfo$Mtd38 = ini*exp((-log(2)/min_half)*(sfo$Days))

sfo = sfo[,c("Days", "Mtd21", "Mtd8", "Mtd38")]
sfo = melt(sfo, id=c("Days"))
ggplot(data=sfo, aes(x=Days, y=value, colour=variable))+
  geom_point() +
  theme_minimal()
```

# Reduce Data Frame

```{r}


dropTheo <- c("No_First", "No_Second" ,"No_Third", "No_Fourth",
               # "CumDays_First", "CumDays_Second", "CumDays_Third", "CumDays_Fourth",
                "remain_1st_29d", "remain_2nd_29d", "remain_3rd_29d", "remain_4th_29d", 
                "remain_1st_46d", "remain_2nd_46d", "remain_3rd_46d", "remain_4th_46d",
                "remain_1st_12d", "remain_2nd_12d", "remain_3rd_12d", "remain_4th_12d",
                "remainTheo.photo" , "remainTheo.photo.prct", "degTheo.photo"
                )

sw <- sw[ , !(names(sw) %in% dropTheo)]

sw <- sw[ 1:(nrow(sw)-1), ]
# Get cummualtive SD
library("TTR")
sw$CumOutSmeto.g.SD <- runSD(sw$TotSMout.g.SD, n=1, cumulative=TRUE)
sw$CumOutMELsm.g.SD <- runSD(sw$MELsm.g.SD, n=1, cumulative=TRUE)
sw <- sw[ 2:nrow(sw), ]

# What does  runSD do? Equal to z :
# x <- df$TotSMout.g.SD
# n <- length(x)
# m <- cumsum(x)/(1:n)
# m1 <- c(NA,m[1:(n-1)])
# ssd <- (x-m)*(x-m1)
# v <- c(0,cumsum(ssd[-1])/(1:(n-1)))
# z <- sqrt(v)

colnames(sw)
```


## Create variables

```{r}

sw$DD.diss <- sw$diss.d13C - d13Co
sw$DD.Bulk <- sw$BulkCatch.d13 - d13Co

# CSIA 
# Elsner degradation B and non-deg fraction f 

# ... for Water
sw$Bw.measure <- 
  (1- ((1000 + d13Co + sw$DD.diss)/(1000+ d13Co))^(1000/epsilon_lab) )*100
sw$Bw.SD1 <-
  (1- ((1000 + d13Co + sw$DD.diss)/(1000+ d13Co))^(1000/epsilon_max) )*100
sw$Bw.SD2 <-
  (1- ((1000 + d13Co + sw$DD.diss)/(1000+ d13Co))^(1000/epsilon_min) )*100



sw$fw.measure <- 100 - sw$Bw.measure
sw$fw.SD1 <- 100 - sw$Bw.SD2
sw$fw.SD2 <- 100 - sw$Bw.SD1

# ... for Soils
sw$Bs.measure <- 
  (1- ((1000 + d13Co + sw$DD.Bulk)/(1000+ d13Co))^(1000/epsilon_lab) )*100
sw$Bs.SD1 <-
  (1- ((1000 + d13Co + sw$DD.Bulk)/(1000+ d13Co))^(1000/epsilon_max) )*100
sw$Bs.SD2 <-
  (1- ((1000 + d13Co + sw$DD.Bulk)/(1000+ d13Co))^(1000/epsilon_min) )*100

# South (maximum fractionation in Abstract)
INSPECT = F
if (INSPECT){
  sw$Bsouth <- 
  (1- ((1000 + d13Co + sw$DD13C.South)/(1000+ d13Co))^(1000/epsilon_lab) )*100

sw$Bsouth.SD1 <- 
  (1- ((1000 + d13Co + sw$DD13C.South)/(1000+ d13Co))^(1000/epsilon_max) )*100
sw$Bsouth.SD2 <- 
  (1- ((1000 + d13Co + sw$DD13C.South)/(1000+ d13Co))^(1000/epsilon_min) )*100
}




sw$fs.measure <- 100 - sw$Bs.measure
sw$fs.SD1 <- 100 - sw$Bs.SD2
sw$fs.SD2 <- 100 - sw$Bs.SD1

# MB
# ... for Water
sw$SMout.measure <- (sw$CumOutSmeto.g/sw$CumAppMass.g)*100
sw$SMout.SD1 <- sw$SMout.measure + (sw$CumOutSmeto.g.SD/sw$CumAppMass.g)*100
sw$SMout.SD2 <- sw$SMout.measure - (sw$CumOutSmeto.g.SD/sw$CumAppMass.g)*100

sw$TPout.measure <- (sw$CumOutMELsm.g/sw$CumAppMass.g)*100
sw$TPout.SD1 <- sw$TPout.measure + (sw$CumOutMELsm.g.SD/sw$CumAppMass.g)*100
sw$TPout.SD2 <- sw$TPout.measure - (sw$CumOutMELsm.g.SD/sw$CumAppMass.g)*100

# ... for soils
#sw$Rem.measure <- (sw$CatchMassSoil.g/sw$CumAppMass.g)*100
#sw$Rem.SD1 <- sw$Rem.measure + (sw$CatchMassSoil.g.SD/sw$CumAppMass.g)*100
#sw$Rem.SD2 <- sw$Rem.measure -  (sw$CatchMassSoil.g.SD/sw$CumAppMass.g)*100

sw$Rem.measure <- (sw$CatchMassSoil.g/sw$CumAppMass.g.OT)*100
sw$Rem.SD1 <- sw$Rem.measure + (sw$CatchMassSoil.g.SD/sw$CumAppMass.g.OT)*100
sw$Rem.SD2 <- sw$Rem.measure -  (sw$CatchMassSoil.g.SD/sw$CumAppMass.g.OT)*100


sw$Unk.measure <- 100 - sw$Rem.measure
sw$Unk.SD1 <- 100 - sw$Rem.SD2
sw$Unk.SD2 <- 100 - sw$Rem.SD1

names(sw)
dropSW <- c("Duration.Hrs" , "TotSMout.g", "TotSMout.g.SD", "MELsm.g" ,"MELsm.g.SD" ,
            "Appl.Mass.g" , "Appl.Mass.g.OT", 
            "CumOutSmeto.g", "CumOutMELsm.g", "ID.N", "BulkCatch.Conc",
             "CumOutSmeto.g.SD",   "CumOutMELsm.g.SD" ,  "DD.diss", "DD.Bulk")

sw <- sw[, !(names(sw) %in% dropSW)]

#sw$LossBulk = 100 - sw$Rem.measure - sw$Bs.measure
#sw$LossBulk.SD1 = 100 - sw$Rem.SD2 - sw$Bs.SD1
#sw$LossBulk.SD2 = 100 - sw$Rem.SD1 - sw$Bs.SD2

```
## Select dates

```{r}

# Delete variables no longer needed
sw$diss.d13C <- NULL
sw$SD.d13C <- NULL
sw$BulkCatch.d13 <- NULL
sw$BulkCatch.d13.SD <- NULL
sw$CatchMassSoil.g <- NULL
sw$CatchMassSoil.g.SD <- NULL
sw$CumAppMass.g <- NULL
sw$CumAppMass.g.OT <- NULL
sw$CumAppMass.g.N.OT <- NULL
sw$CumAppMass.g.T.OT <- NULL
sw$CumAppMass.g.S.OT <- NULL

April <- subset(sw, (Date.ti >= as.POSIXct("2016-04-18 20:30:00", tz = "EST") 
                 & Date.ti < as.POSIXct("2016-05-01 00:00:00", tz = "EST"))) 
                     # & Date.ti <= as.POSIXct("2016-05-03 13:11:00", tz = "EST"))) 

#soils.April <- soils.April[complete.cases(soils.April[ , "ID.N"]), ]

June <- subset(sw, (Date.ti >= as.POSIXct("2016-06-01 00:56:00", tz = "EST") 
                 & Date.ti <= as.POSIXct("2016-06-28 14:52:00", tz = "EST")) )
# Narrow range
June <- subset(sw, (Date.ti >= as.POSIXct("2016-06-07 12:00:00", tz = "EST") 
                 & Date.ti <= as.POSIXct("2016-06-14 13:06:00", tz = "EST")) )
#soils.June <- soils.June[complete.cases(soils.June[ , "ID.N"]), ]


# Take an average of the obtained values for selected date ranges
library(dplyr)
a <- April %>% 
  summarise_each(funs(mean(. , na.rm=T) )) 

j <- June %>% 
  summarise_each(funs(mean(. , na.rm=T) )) 

a$Month <- "April"
j$Month <- "June"

```



## Tidying up and plotting

```{r}
# Get rid of date and cummulative days, and merge april and june
aj <- rbind(a, j)[, 6:ncol(j) ]
ajTidy <- aj %>%
  gather(measure, value, -Month) %>% # Melts data frame
  separate(measure, into = c("Sink", "temporary_var")) %>% # parses the sep = "." into...
  spread(temporary_var, value) # Moves molten temporary variable to own column  

type <- rep(c("CSIA - Soils", "CSIA - Outlet", 
              "Predicted",  
              "CSIA - Soils", "CSIA - Outlet", 
              "MB - Soils", 
              "Predicted", 
              "MB - Outlet", "MB - Outlet", "MB - Soils"), 2)

ajTidy <- cbind(ajTidy, type)
ajTidy$Sink <- as.factor(ajTidy$Sink)

ajTidy$measure <- as.numeric(ajTidy$measure)
ajTidy$SD1 <- as.numeric(ajTidy$SD1)
ajTidy$SD2 <- as.numeric(ajTidy$SD2)

levels(ajTidy$Sink)
ajTidy$Sink <- factor(ajTidy$Sink, levels = c("Bs" , "fs" , "Unk", "Rem", 
                                              "Bw", "TPout" , "fw" , "SMout",
                                              "degTheo", "remainTheo"))


write.csv2(ajTidy, "Data/BarPlotData_R.csv", row.names = F) 
```

## Plot Soils

```{r}
SoilBars <- ggplot(data = subset(ajTidy, 
                                 Sink == "Bs" | Sink == "fs"
                                 | Sink == "Rem" | Sink == "Unk"
                                 ) , aes(x=Month, y=measure, fill = Sink, ymin=SD1, ymax=SD2))+
  geom_bar(stat = "identity", position = "dodge", width = NULL) +
  geom_errorbar(#aes(ymin=SD1, ymax=SD2),
                  position=position_dodge(0.9),
                  width=.4  # ) + #,                    # Width of the error bars
                  ) + 
  theme_bw() +
  ylab("% S-met Applied") +
  theme(text = element_text(size=15),
        axis.title.x = element_blank(),
        aspect.ratio = 1
        ) +
  xlab("Month") +
  facet_wrap(~type) +
  scale_y_continuous( breaks=c(25, 50, 75, 100), limits = c(0, 100) )+ #expand=c(0, 10, 0, 0)) + 
  scale_fill_manual(#values = c("#6a51a3" , "#ec7014", "#807dba", "#fe9929"), # purple-orange
                    values = c("#6a51a3" , "#ec7014", "#d9d9d9", "#fe9929"), # Unknown as grey
                    name= "Top Soils (A, B)" ,# element_blank(), #"Mass Balance", # \n
                    breaks=c("Bs", "fs" , 
                                "Unk" , "Rem" 
                                ),
                    labels=c("Degr.", "Non-degr.", 
                                "Unknown", "Persist." ))+
  guides(fill=guide_legend(ncol=2))

SoilBars

ajTidy$CI95 <- signif(ajTidy$measure - ajTidy$SD1, 2)

# ggsave(SoilBars, filename = "images/SoilBars.png") 
```

## Plot Waters

```{r}

OutBars <- ggplot(data = subset(ajTidy, 
                                type != 'Predicted' 
                                & Sink != "Bs" 
                                &  Sink != "fs"
                                & Sink != "Rem"
                                & Sink != "Unk"
                                ) , 
                  aes(x=Month, y=measure, fill = Sink, ymin=SD1, ymax=SD2))+
  geom_bar(stat = "identity", position = "dodge", width = NULL) +
  geom_errorbar(#aes(ymin=SD1, ymax=SD2),
                  position=position_dodge(0.9),
                  width=.4
                  ) + 
  theme_bw() +
  ylab("% S-met Applied") +
  theme(text = element_text(size=15),
        axis.title.x = element_blank(),
        aspect.ratio = 1) +
  scale_y_continuous( breaks=c(25, 50, 75, 100), limits = c(0, 100) )+ #expand=c(0, 10, 0, 0)) + 
  #xlab("Month") +
  facet_wrap(~type) +
  scale_fill_manual(#values = c("#01665e",  "#ec7014",   "#35978f",  "#fe9929", "#80cdc1", "#fec44f"), # blue-orange
                    values = c("#01665e",  "#35978f", "#ec7014" ,  "#fe9929"), # blue-orange
                    #values = c("#238b45", "#41ab5d", "#74c476", "#40004b", "#762a83", "#9970ab" ), # green-purple
                    #values = c("#238b45", "#41ab5d", "#74c476", "#ec7014", "#fe9929", "#fec44f" ), # green-orange
                    #values = c("#80cdc1", "#018571", "#a6611a", "#dfc27d", "#80cdc1", "#018571"),
                    name= "Outlet (C, D)" ,# element_blank(), #"Mass Balance", # \n
                       breaks=c("Bw", "fw" , 
                                "TPout" , "SMout" #, 
                                #"Deg", "Rem" 
                                ),
                       labels=c("Degr.", "Non-degr.", 
                                "TPs loads", "S-met loads" #, 
                                #"Degr." , "Persist."
                                 )) +
  guides(fill=guide_legend(ncol=3))


OutBars
# ggsave(OutBars, filename = "images/OutBars.png") 
```

## Theoretical

```{r}

## Predicted (half-life) only
theoBars <- ggplot(data = subset(ajTidy, type == 'Predicted') , 
                  aes(x=Month, y=measure, fill = Sink, ymin=SD1, ymax=SD2))+
  geom_bar(stat = "identity", position = "dodge", width = NULL) +
  geom_errorbar(#aes(ymin=SD1, ymax=SD2),
                 position=position_dodge(0.9),
                 width=.4
                 ) + 
  theme_bw() +
  # ylab("% S-met Applied") +
  theme(text = element_text(size=15),
        axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        # axis.text.y = element_blank(),
        aspect.ratio = 1) +
  scale_y_continuous( breaks=c(25, 50, 75, 100), limits = c(0, 100) )+ #expand=c(0, 10, 0, 0)) +

  #xlab("Month") +
  facet_wrap(~type) +
  scale_fill_manual(values = c( "#80cdc1", "#fec44f"), # blue-orange
                    #values = c("#238b45", "#41ab5d", "#74c476", "#40004b", "#762a83", "#9970ab" ), # green-purple
                    #values = c("#238b45", "#41ab5d", "#74c476", "#ec7014", "#fe9929", "#fec44f" ), # green-orange
                    #values = c("#80cdc1", "#018571", "#a6611a", "#dfc27d", "#80cdc1", "#018571"),
                    name=expression("Predicted, " ~ t[1/2] ~ "(E)"),
                    # name= "Predicted (half-life) (E)",# element_blank(), #"Mass Balance", # \n
                    breaks=c("degTheo", "remainTheo"),
                    labels=c("Degr." , "Persist.")
                    
                    ) +
  guides(fill=guide_legend(ncol=1))

#pal = c( "#35978f", "#fe9929", "#fec44f", "#01665e", "#80cdc1", "#ec7014" )
#display.pal(pal, sel=1:length(pal), names=F)

theoBars
# ggsave(theoBars, filename = "images/TheoBars.png") 
```

## Merge both Outlet and Soils - BARS

```{r}

#balAll <- rbind(balTidyType,  balTidySol)

OutBars_noLeg <- OutBars + theme(legend.position = 'none')
OutBars_Leg <- get_legend(OutBars)

SoilBars_noLeg <- SoilBars + theme(legend.position = 'none')
SoilBars_Leg <- get_legend(SoilBars)

TheoBars_noLeg <- theoBars + theme(legend.position = 'none')
TheoBars_Leg <- get_legend(theoBars)

#plot_grid(OutBars_noLeg, SoilBars_noLeg,
#                    ncol =1, nrow = 2, align ="v" )
#,
#                    labels = c("A", "C", "B", "D"))

balAllplot <- ggdraw() +
  
  draw_plot(SoilBars_noLeg, x=-.2, y = 0.55, width = 1, height = .4) +
  draw_plot(TheoBars_noLeg, x=0.23, y=.55, width =  1, height = .4) +
  draw_plot(OutBars_noLeg, x=-.2, y=.05, width =  1, height = .4) +
  
  draw_label("A", x= 0.15, y = .86, size = 12, fontface = "bold") +
  draw_label("C", x= 0.15, y = .35, size = 12, fontface = "bold") +
  draw_label("B", x= 0.39, y = .86, size = 12, fontface = "bold") +
  draw_label("D", x= 0.39, y = .35, size = 12, fontface = "bold") +
  draw_label("E", x= 0.67, y = .86, size = 12, fontface = "bold") +
  
  draw_plot(SoilBars_Leg, x=(0.243 + 0.05) , y = 0.35, width = 1, height = 0.1) +
  draw_plot(OutBars_Leg, x=(0.280 + 0.05), y = 0.2, width = 1, height = 0.1) +
  draw_plot(TheoBars_Leg, x=(0.217 + 0.05) , y = 0.05, width = 1, height = 0.1) 

balAllplot

#balAllplot <- ggdraw() +
#  draw_plot(SoilBars_noLeg, x=0.0, y = 0.5, width = 0.5, height = 0.55) +
#  draw_plot(TheoBars_noLeg, x=0.65, y=.5, width =  0.33, height = .5) + 
#  draw_plot(OutBars_noLeg, x=0.01, y=.0, width =  0.60, height = .5) + 
#  draw_plot(SoilBars_Leg, x=0.67, y = 0.3, width = 0.1, height = 0.1) +
#  draw_plot(OutBars_Leg, x=0.695, y = 0.1, width = 0.1, height = 0.1) +
#  draw_plot(TheoBars_Leg, x=0.89, y = 0.3, width = 0.05, height = 0.1) +
#  draw_label("A", x= 0.11, y = .9, size = 12, fontface = "bold") +
#  draw_label("C", x= 0.11, y = .39, size = 12, fontface = "bold") +
#  draw_label("B", x= 0.37, y = .9, size = 12, fontface = "bold") +
#  draw_label("D", x= 0.37, y = .39, size = 12, fontface = "bold") +
#  draw_label("E", x= 0.75, y = .9, size = 12, fontface = "bold") 

# balAllplot
#ggsave(balAllplot, filename = "images/MB_CSIA_Bars.png", width = 8, height = 5, units = "in", scale = 1)


SAVE = F
PC = F
if (SAVE){
  if (PC){
      ggsave(balAllplot, 
           filename = "D:/Documents/these_pablo/WriteUp/Alteck_PNAS_LaTeX/images/balAllplot2.pdf", 
           device = "pdf", dpi = 600,  scale = 1, # )# ,
           width = 7, height = 5.5)
    } else {
      ggsave(balAllplot, 
         filename = "/Users/DayTightChunks/Documents/PhD/Writeups/PNAS/Alteck_PNAS_LaTeX/images/balAllplot.pdf", 
         device="pdf", dpi = 600,  scale = 1, # )# ,
         width = 7, height = 5.5)
    }
  }


```


# Simplified Figure

```{r}
ajTidy$Sink = as.character(ajTidy$Sink)
simp = subset(ajTidy, Sink == "remainTheo" | Sink == "fs" | Sink == "fw" | Sink == "Rem" | Sink == "TPout")

simp$Sink = as.factor(simp$Sink)
levels(simp$Sink)
simp$Sink <- factor(simp$Sink, levels = c("remainTheo", "fs", "Rem", "fw", "TPout"))

red = "#F8766D"
green = "#00BA38"
brown = "#DE8C00"
violet = "#C77CFF"
palegreen = "#B79F00"
darkblue = "#619CFF"
brightblue1 = "#00BFC4"
brightblue2 = "#00B4F0"

simplebar = ggplot(data = simp,  aes(x=Month, y=measure, fill = Sink, ymin=SD1, ymax=SD2))+
  geom_bar(stat = "identity", position = "dodge", width = NULL) +
  geom_errorbar(#aes(ymin=SD1, ymax=SD2),
                 position=position_dodge(0.9),
                 width=.4
                 ) + 
  theme_minimal() +
  ylab("% S-met") +
  theme(text = element_text(size=17),
        axis.title.x = element_blank()) +
        scale_fill_manual(name= "Estimation Approach",
                          values = c(violet , green, brown, darkblue, brightblue1),
                          breaks=c("remainTheo", "fs", "Rem", "fw", "TPout"),
                          labels=c("A. Reference Non-degr. (DT50)",  "B. Soil Non-degr. (CSIA-f)", "C. Soil Remain. Mass (MB)", 
                                   "D. Outlet Non-degr. (CSIA-f)", "E. Outlet Loadings (MB)" )) +
        guides(fill=guide_legend(ncol=1))
  
simplebar


SAVE = F
PC = T
if (SAVE){
  if (PC){
      ggsave(simplebar, 
           filename = "D:/Documents/these_pablo/WriteUp/Images/barplot.pdf", 
           device = "pdf", dpi = 600,  scale = 1, # )# ,
           width = 7, height = 5)
    } else {
      ggsave(simplebar, 
         filename = "/Users/DayTightChunks/Documents/PhD/Writeups/PNAS/Alteck_PNAS_LaTeX/images/balAllplot.pdf", 
         device="pdf", dpi = 600,  scale = 1, # )# ,
         width = 7, height = 5)
    }
  }

```

# Color pallette

```{r}
library(scales) 
## Color palette
show_col(hue_pal()(12))



```
