==> 12.0_Presentation.Rmd <==
---
title: "Presentation Figures"
author: "PAZ"
date: "05/10/2017"
output: pdf_document
---

```{r, echo=FALSE, message=FALSE, include=FALSE} 
knitr::opts_chunk$set(echo = FALSE, message=FALSE, warning = F)

Sys.setlocale("LC_ALL", "English") 

library("tidyr")
#library("dplyr")
library("ggplot2")
library("ggrepel")
library("zoo")
library("scales")

# Import lab values 
source("global.R")
```

```{r}
# Import raw data
enrich = read.csv2('Data/EnrichmentExp.csv', sep = ";", dec = ".", header = T)
# enrich = subset(enrich, C.SM < 3)
df = subset(enrich, Temp != 30)
df$DD13 <- df$Delta - delta_ini

bio = subset(enrich, Type==as.character("Biotic") & Temp == 20)
abiotic = subset(enrich, Type==as.character("Abiotic") & Temp == 20)

bio$DD13 <- bio$Delta - delta_ini
pearson_r_bio <- cor.test(bio$Delta , bio$C.SM)[4]
r_label_bio <- sprintf("Pearson~r == %0.2f", pearson_r_bio)
p_value_bio <- cor.test(bio$Delta, bio$C.SM)[3]

if (p_value_bio < 0.0001){
  p_label_bio <- "(P < 0.001)"
} else if (p_value_bio < 0.001) {
  p_label_bio <- "(P < 0.001)"
} else if (p_value_bio < 0.015) {
  p_label_bio <- ("P < 0.01")
} else {
  p_label_bio <-  sprintf("P == %0.2f", p_value_bio)
}

abiotic$DD13 <- abiotic$Delta - delta_ini
pearson_r_abi <- cor.test(abiotic$Delta , abiotic$C.SM)[4]
r_label_abi <- sprintf("Pearson~r == %0.2f", pearson_r_abi)
p_value_abi <- cor.test(abiotic$Delta, abiotic$C.SM)[3]

if (p_value_abi < 0.0001){
  p_label_abi <- "(P < 0.001)"
} else if (p_value_abi < 0.001) {
  p_label_abi <- "(P < 0.001)"
} else if (p_value_abi < 0.015) {
  p_label_abi <- ("P < 0.01")
} else {
  p_label_abi <-  sprintf("P == %0.2f", p_value_abi)
}

```

## Plot ABiotic and Biotic together

```{r}
names(df) <- c("Type",
               "Experiment",   
               "Temp",
               "Moisture",
               "Days",
               "Conc.measure", 
               "Delta.measure",
               "Delta.SD",
               "DD13.measure")
df$Moisture = ifelse(df$Moisture == 20, 0.2, 0.4)
df$Moisture <- as.factor(df$Moisture)
df$DD13.SD = df$Delta.SD
df$Experiment = ifelse(df$Experiment == "CA", "Abiotic", "Biotic")
molten <- df %>%
  gather(measure, value, -Type,
               -Experiment, -Temp, -Moisture, -Days) %>% # Melts data frame
  separate(measure, into = c("Location", "temporary_var")) %>% # parses the sep = "." into...
  # Location will be first string of variable name
  spread(temporary_var, value)


# library(scales) 
## Color palette
# show_col(hue_pal()(12))
# Red: #F8766D
# Green: #00BA38
# Blue: #619CFF
# Default blue: #00BFC4
```

## Biotic

```{r}
b = ggplot(df, aes(x=Conc.measure, y = DD13.measure, group = Experiment, colour = Experiment)) +
  theme_minimal() +
  geom_errorbar(data= subset(df, Experiment =="Biotic"), 
                aes(ymin = DD13.measure - DD13.SD, ymax = DD13.measure + DD13.SD)) +
  # geom_point(data= subset(df, Experiment !="Biotic"), aes(size = Days)) +
  geom_point(data= subset(df, Experiment =="Biotic"), aes(size = Days, shape = Moisture)) +
  # stat_smooth(data = subset(df,  Experiment !="Biotic"),
  #            aes(x=Conc.measure, y=DD13.measure), method = "lm", formula = y ~ x, se=F) +
  stat_smooth(data = subset(df,  Experiment =="Biotic"),
              aes(x=Conc.measure, y=DD13.measure), method = "lm", formula = y ~ x, se=F) +
  scale_size_continuous(range = c(1, 5), breaks= c(1, 10, 50, 100, 200), limits = c(1, 200)) + 
  scale_y_continuous(breaks=c(0, 1, 2, 3 , 4 ,5, 6, 7 , 8) ) + 
  theme(# legend.position = "top" ,
        text = element_text(size=23)) +
  # labs(size="   Days after spiking") + #, shape = "Mass Carbon") +
  ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  xlab(expression(paste("S-met Soil Concentration  ", {({mu}*g / g~dry~wt.)}))) +
  #annotate("text", x = 1.5, y = 6.7, label = as.character(r_label_bio), colour = "#00BFC4", size = 5, parse = T) + 
  #annotate("text", x = 1.5, y = 6.0, label = p_label_bio, parse = T, size = 5, colour = "#00BFC4") +
  annotate("text", x = 1.5, y = 6.7, label = as.character(r_label_bio), colour = "#F8766D", size = 5, parse = T) + 
  annotate("text", x = 1.5, y = 6.0, label = p_label_bio, parse = T, size = 5, colour = "#F8766D") +
  annotate("rect", xmin=0, xmax=max(df$Conc.measure), ymin=-0.8, ymax=0.8, alpha=0.2)

b



# ggsave(b, filename = "images/biotic_soils.png" , width = 8.7, height = 6 )#, units = "cm", scale = 1)


bio$yRaleigh <- log((1000+delta_ini+bio$DD13)/(1000+delta_ini))
bio$xRaleigh <- log(bio$C.SM/c_ini)
expModel<-lm(yRaleigh~xRaleigh, data= bio) 

cofsoil <- as.numeric(coef(expModel)[2]*1000)
minX <- confint(expModel, "xRaleigh", level = 0.95)[1]*1000
maxX <- confint(expModel, "xRaleigh", level = 0.95)[2]*1000
se <- summary(expModel)$coef[[4]]*1000

e_label <- sprintf("%0.2f", cofsoil)

e_label2 <- substitute(paste(epsilon ["lab"] , " = ", epsilon_f, " \u00B1 ", "0.47" ,"\u2030"),
                     list(epsilon_f = signif(cofsoil, 3)))

cond_label = paste("theta:~20~+~40~and~degree~C:20")

CI95 = maxX - cofsoil

bray = ggplot(data = subset(bio,  !is.na(yRaleigh) ), aes(x=xRaleigh, y=yRaleigh)) +
  geom_point() +
  theme_minimal() +
  theme(text = element_text(size=23), 
        legend.position = "top" 
        ) +
  ylab(expression(paste("ln(Rt/R0)"))) +
  xlab(expression(paste("ln([Conc]t/[Conc]0)"))) +
  
  stat_smooth(method = "lm", formula = y ~ x, se=F) 
  #annotate("text", x = -1.5, y = 0.007, label = as.character(e_label), parse = T, size = 3.5) +
  #annotate("text", x = -1.5, y = 0.006, label = e_label2, parse = T, size = 3.5) +
  # annotate("text", x = -1.5, y = 0.006, label = cond_label, parse = T, size = 3.5) +
  

gbray = ggdraw() + 
   draw_plot(bray, x=0, y = 0.0, width = 1, height = 1) +
   draw_label(e_label2 , x= .4, y = .30, size = 20)

#ggsave(gbray, filename = "images/biotic_rayleigh.png" , width = 8.7, height = 6 )#, units = "cm", scale = 1)

summary(expModel)
```

## Abiotic

```{r}
ab = ggplot(df, aes(x=Conc.measure, y = DD13.measure, group = Experiment, colour = Experiment)) +
  theme_minimal() +
  geom_errorbar(data= subset(df, Experiment =="Abiotic"), 
                aes(ymin = DD13.measure - DD13.SD, ymax = DD13.measure + DD13.SD)) +
  # geom_point(data= subset(df, Experiment !="Abiotic"), aes(size = Days)) +
  geom_point(data= subset(df, Experiment =="Abiotic"), aes(size = Days, shape = Moisture)) +
  # stat_smooth(data = subset(df,  Experiment !="Abiotic"),
  #            aes(x=Conc.measure, y=DD13.measure), method = "lm", formula = y ~ x, se=F) +
  stat_smooth(data = subset(df,  Experiment =="Abiotic"),
              aes(x=Conc.measure, y=DD13.measure), method = "lm", formula = y ~ x, se=F) +
  scale_size_continuous(range = c(1, 5), breaks= c(1, 10, 50, 100, 200), limits = c(1, 200)) + 
  scale_y_continuous(breaks=c(0, 1, 2, 3 , 4 ,5, 6, 7 , 8) ) + 
  theme(# legend.position = "top" ,
        text = element_text(size=23)) +
  # labs(size="   Days after spiking") + #, shape = "Mass Carbon") +
  ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  xlab(expression(paste("S-met Soil Concentration  ", {({mu}*g / g~dry~wt.)}))) +
  #annotate("text", x = 1.5, y = 6.7, label = as.character(r_label_bio), colour = "#00BFC4", size = 5, parse = T) + 
  #annotate("text", x = 1.5, y = 6.0, label = p_label_bio, parse = T, size = 5, colour = "#00BFC4") +
  annotate("text", x = 1.5, y = 6.7, label = as.character(r_label_abi), colour = "#F8766D", size = 5, parse = T) + 
  annotate("text", x = 1.5, y = 6.0, label = p_label_abi, parse = T, size = 5, colour = "#F8766D") +
  annotate("rect", xmin=0, xmax=max(df$Conc.measure), ymin=-propagatedError, ymax=propagatedError, alpha=0.2)

ab

# ggsave(ab, filename = "images/abiotic_soils.png" , width = 8.7, height = 6 )#, units = "cm", scale = 1)
```

## Both experiments together

```{r}
mix = ggplot(df, aes(x=Conc.measure, y = DD13.measure, group = Experiment, colour = Experiment)) +
  theme_minimal() +
  geom_errorbar(aes(ymin = DD13.measure - DD13.SD, ymax = DD13.measure + DD13.SD)) +
  geom_point(data= subset(df, Experiment !="Abiotic"), aes(size = Days, shape = Moisture)) +
  geom_point(data= subset(df, Experiment =="Abiotic"), aes(size = Days, shape = Moisture)) +
  stat_smooth(data = subset(df,  Experiment !="Abiotic"),
              aes(x=Conc.measure, y=DD13.measure), method = "lm", formula = y ~ x, se=F) +
  stat_smooth(data = subset(df,  Experiment =="Abiotic"),
              aes(x=Conc.measure, y=DD13.measure), method = "lm", formula = y ~ x, se=F) +
  scale_size_continuous(range = c(1, 5), breaks= c(1, 10, 50, 100, 200), limits = c(1, 200)) + 
  scale_y_continuous(breaks=c(0, 1, 2, 3 , 4 ,5, 6, 7 , 8) ) + 
  # geom_errorbarh(aes(xmin = Conc.mug.g.dry.soil - Conc.ComSoil.SD, xmax = Conc.mug.g.dry.soil + Conc.ComSoil.SD)) +
  theme(# legend.position = "top" ,
        text = element_text(size=23)) +
  # labs(size="   Days after spiking") + #, shape = "Mass Carbon") +
  ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  xlab(expression(paste("S-met Soil Concentration  ", {({mu}*g / g~dry~wt.)}))) +
  annotate("text", x = 1.5, y = 6.7, label = as.character(r_label_bio), colour = "#00BFC4", size = 5, parse = T) + 
  annotate("text", x = 1.5, y = 6.0, label = p_label_bio, parse = T, size = 5, colour = "#00BFC4") +
  annotate("text", x = 1.5, y = 4.7, label = as.character(r_label_abi), colour = "#F8766D", size = 5, parse = T) + 
  annotate("text", x = 1.5, y = 4.0, label = p_label_abi, parse = T, size = 5, colour = "#F8766D") +
  annotate("rect", xmin=0, xmax=max(df$Conc.measure), ymin=-propagatedError, ymax=propagatedError, alpha=0.2)

mix
ggsave(mix, filename = "images/biotic_soils.png" , width = 8.7, height = 6 )#, units = "cm", scale = 1)
```

## Bar plots

```{r}
library(dplyr)
library(tidyr)
library(zoo)
library(reshape)
library(ggplot2)
library("ggrepel")

library("plotly")
library("cowplot")
library("gridExtra")
library("Cairo")
library("GGally")
library("scales")
ajTidy = read.csv2('Data/BarPlotData_R.csv') # , sep = ";", dec = ".", header = T)
ajTidy$measure <- as.numeric(ajTidy$measure)
ajTidy$SD1 <- as.numeric(ajTidy$SD1)
ajTidy$SD2 <- as.numeric(ajTidy$SD2)

ajTidy$Sink <- factor(ajTidy$Sink, levels = c("Bs" , "fs" , "Unk", "Rem", 
                                              "Bw", "TPout" , "fw" , "SMout",
                                              "degTheo", "remainTheo"))
```

## Plot Soils

```{r}
SoilBars <- ggplot(data = subset(ajTidy, 
                                 Sink == "Bs" | Sink == "fs"
                                 | Sink == "Rem" | Sink == "Unk" 
                                   # Sink == "degTheo" | Sink == "remainTheo"
                                 ) , aes(x=Month, y=measure, fill = Sink, ymin=SD1, ymax=SD2))+
  geom_bar(stat = "identity", position = "dodge", width = NULL) +
  geom_errorbar(#aes(ymin=SD1, ymax=SD2),
                  position=position_dodge(0.9),
                  width=.4  # ) + #,                    # Width of the error bars
                  ) + 
  theme_bw() +
  ylab("% S-met Applied") +
  theme(text = element_text(size=15),
        axis.title.x = element_blank(),
        aspect.ratio = 1
        ) +
  xlab("Month") +
  facet_wrap(~type) +
  scale_y_continuous( breaks=c(25, 50, 75, 100), limits = c(0, 100) )+ #expand=c(0, 10, 0, 0)) + 
  scale_fill_manual(#values = c("#6a51a3" , "#ec7014", "#807dba", "#fe9929"), # purple-orange
                    values = c("#6a51a3" , "#ec7014", "#d9d9d9", "#fe9929"), # Unknown as grey
                    name= "Top Soils (A, B)" ,# element_blank(), #"Mass Balance", # \n
                    breaks=c("Bs", "fs" , 
                                "Unk" , "Rem"
                             
                                ),
                    labels=c("Degr.", "Non-degr.", 
                                "Unknown", "Persist." ))+
  guides(fill=guide_legend(ncol=2))

SoilBars

# ajTidy$CI95 <- signif(ajTidy$measure - ajTidy$SD1, 2)

# ggsave(SoilBars, filename = "images/SoilBars.png") 
```

## Plot Waters

```{r}

OutBars <- ggplot(data = subset(ajTidy, 
                                type != 'Predicted' 
                                & Sink != "Bs" 
                                &  Sink != "fs"
                                & Sink != "Rem"
                                & Sink != "Unk"
                                ) , 
                  aes(x=Month, y=measure, fill = Sink, ymin=SD1, ymax=SD2))+
  geom_bar(stat = "identity", position = "dodge", width = NULL) +
  geom_errorbar(#aes(ymin=SD1, ymax=SD2),
                  position=position_dodge(0.9),
                  width=.4
                  ) + 
  theme_bw() +
  ylab("% S-met Applied") +
  theme(text = element_text(size=15),
        axis.title.x = element_blank(),
        aspect.ratio = 1) +
  scale_y_continuous( breaks=c(25, 50, 75, 100), limits = c(0, 100) )+ #expand=c(0, 10, 0, 0)) + 
  #xlab("Month") +
  facet_wrap(~type) +
  scale_fill_manual(#values = c("#01665e",  "#ec7014",   "#35978f",  "#fe9929", "#80cdc1", "#fec44f"), # blue-orange
                    values = c("#01665e",  "#35978f", "#ec7014" ,  "#fe9929"), # blue-orange
                    #values = c("#238b45", "#41ab5d", "#74c476", "#40004b", "#762a83", "#9970ab" ), # green-purple
                    #values = c("#238b45", "#41ab5d", "#74c476", "#ec7014", "#fe9929", "#fec44f" ), # green-orange
                    #values = c("#80cdc1", "#018571", "#a6611a", "#dfc27d", "#80cdc1", "#018571"),
                    name= "Outlet (C, D)" ,# element_blank(), #"Mass Balance", # \n
                       breaks=c("Bw", "fw" , 
                                "TPout" , "SMout" #, 
                                #"Deg", "Rem" 
                                ),
                       labels=c("Degr.", "Non-degr.", 
                                "TPs loads", "S-met loads" #, 
                                #"Degr." , "Persist."
                                 )) +
  guides(fill=guide_legend(ncol=3))


OutBars
# ggsave(OutBars, filename = "images/OutBars.png") 
```

## Theoretical

```{r}

## Predicted (half-life) only
theoBars <- ggplot(data = subset(ajTidy, type == 'Predicted') , 
                  aes(x=Month, y=measure, fill = Sink, ymin=SD1, ymax=SD2))+
  geom_bar(stat = "identity", position = "dodge", width = NULL) +
  geom_errorbar(#aes(ymin=SD1, ymax=SD2),
                 position=position_dodge(0.9),
                 width=.4
                 ) + 
  theme_bw() +
  # ylab("% S-met Applied") +
  theme(text = element_text(size=15),
        axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        # axis.text.y = element_blank(),
        aspect.ratio = 1) +
  scale_y_continuous( breaks=c(25, 50, 75, 100), limits = c(0, 100) )+ #expand=c(0, 10, 0, 0)) +

  #xlab("Month") +
  facet_wrap(~type) +
  scale_fill_manual(values = c( "#80cdc1", "#fec44f"), # blue-orange
                    #values = c("#238b45", "#41ab5d", "#74c476", "#40004b", "#762a83", "#9970ab" ), # green-purple
                    #values = c("#238b45", "#41ab5d", "#74c476", "#ec7014", "#fe9929", "#fec44f" ), # green-orange
                    #values = c("#80cdc1", "#018571", "#a6611a", "#dfc27d", "#80cdc1", "#018571"),
                    name=expression("Predicted, " ~ t[1/2] ~ "(E)"),
                    # name= "Predicted (half-life) (E)",# element_blank(), #"Mass Balance", # \n
                    breaks=c("degTheo", "remainTheo"),
                    labels=c("Degr." , "Persist.")
                    
                    ) +
  guides(fill=guide_legend(ncol=1))

#pal = c( "#35978f", "#fe9929", "#fec44f", "#01665e", "#80cdc1", "#ec7014" )
#display.pal(pal, sel=1:length(pal), names=F)

theoBars
# ggsave(theoBars, filename = "images/TheoBars.png") 
```

## Merge both Outlet and Soils - BARS

```{r}

#balAll <- rbind(balTidyType,  balTidySol)

OutBars_noLeg <- OutBars + theme(legend.position = 'none')
OutBars_Leg <- get_legend(OutBars)

SoilBars_noLeg <- SoilBars + theme(legend.position = 'none')
SoilBars_Leg <- get_legend(SoilBars)

TheoBars_noLeg <- theoBars + theme(legend.position = 'none')
TheoBars_Leg <- get_legend(theoBars)

#plot_grid(OutBars_noLeg, SoilBars_noLeg,
#                    ncol =1, nrow = 2, align ="v" )
#,
#                    labels = c("A", "C", "B", "D"))

balAllplot <- ggdraw() +
  
  draw_plot(SoilBars_noLeg, x=-.2, y = 0.55, width = 1, height = .4) +
  draw_plot(TheoBars_noLeg, x=0.23, y=.55, width =  1, height = .4) +
  draw_plot(OutBars_noLeg, x=-.2, y=.05, width =  1, height = .4) +
  
  draw_label("A", x= 0.15, y = .85, size = 12, fontface = "bold") +
  draw_label("C", x= 0.15, y = .35, size = 12, fontface = "bold") +
  draw_label("B", x= 0.39, y = .85, size = 12, fontface = "bold") +
  draw_label("D", x= 0.39, y = .35, size = 12, fontface = "bold") +
  draw_label("E", x= 0.67, y = .85, size = 12, fontface = "bold") +
  
  draw_plot(SoilBars_Leg, x=(0.243 + 0.05) , y = 0.35, width = 1, height = 0.1) +
  draw_plot(OutBars_Leg, x=(0.280 + 0.05), y = 0.2, width = 1, height = 0.1) +
  draw_plot(TheoBars_Leg, x=(0.217 + 0.05) , y = 0.05, width = 1, height = 0.1) 

balAllplot

#balAllplot <- ggdraw() +
#  draw_plot(SoilBars_noLeg, x=0.0, y = 0.5, width = 0.5, height = 0.55) +
#  draw_plot(TheoBars_noLeg, x=0.65, y=.5, width =  0.33, height = .5) + 
#  draw_plot(OutBars_noLeg, x=0.01, y=.0, width =  0.60, height = .5) # +
  
# draw_plot(SoilBars_Leg, x=0.67, y = 0.3, width = 0.1, height = 0.1) +
#  draw_plot(OutBars_Leg, x=0.695, y = 0.1, width = 0.1, height = 0.1) +
#  draw_plot(TheoBars_Leg, x=0.89, y = 0.3, width = 0.05, height = 0.1) #+
#  draw_label("A", x= 0.11, y = .9, size = 12, fontface = "bold") +
 # draw_label("C", x= 0.11, y = .39, size = 12, fontface = "bold") +
  #draw_label("B", x= 0.37, y = .9, size = 12, fontface = "bold") +
  #draw_label("D", x= 0.37, y = .39, size = 12, fontface = "bold") +
  #draw_label("E", x= 0.75, y = .9, size = 12, fontface = "bold") 

# balAllplot
# ggsave(balAllplot, filename = "images/Bars.png", width = 8, height = 5 ) #, units = "in", scale = 1)
```

## Outlet and Soils vs Time

```{r}
wstidier2 = read.csv2('Data/OutletSoilsPlot_R.csv')
wstidier2$Date <- as.POSIXct(strptime(wstidier2$Date, 
                                          "%Y-%m-%d %H:%M", tz="EST")) 

```

# All Data

## Lab Enrichment

```{r}
mindate = min(wstidier2$Date)
maxdate = max(wstidier2$Date)

pd <- position_dodge(width = 0.5)
limits_DdC <- aes(ymin=measure-SD, ymax=measure+SD, colour = Source)

wsALL_lab <- ggplot(data = wstidier2, aes(x = Date, y = measure, group = Source) )+ 
  geom_errorbar(data=subset(wstidier2, Source == 'Bulk'), limits_DdC, size=0.2) +
  geom_errorbar(data=subset(wstidier2, Source == 'South' 
                            | Source == 'North'
                            | Source == 'Valley'
                            ), limits_DdC, size=0.1) +
  geom_errorbar(data=subset(wstidier2, Source == 'Outlet'), limits_DdC) +
  geom_point(data=subset(wstidier2, (Source == 'South'
                                     | Source == 'North'
                                     | Source == 'Valley'
                                     ) 
                         # & Date > as.POSIXct('2016-05-14 08:04:00')
                         ), 
             aes(shape = Type, 
                 colour = Source)) +
  geom_point(data=subset(wstidier2, Source == 'Outlet'), aes(shape = Type, colour = Source, size = Qmean)) +
  geom_point(data=subset(wstidier2, Source == 'Bulk'), aes(shape = Type, colour = Source)) +
  
  # Water
  stat_smooth(data=subset(wstidier2,
                          (Source == "Outlet" 
                           # & Event > 1 
                           & Type == "Dissolved (Outlet)")), 
              method = "lm", formula = y ~ poly(x, 2), se = F, aes(colour = 'Outlet'), alpha = 0.2, size=0.2) +
  # North
  stat_smooth(data=subset(wstidier2,
                          (Source == "Bulk" )), #| Source == "South" )), 
              method = "lm", formula = y ~ poly(x, 2), se = F, aes(colour = 'Bulk'), alpha = 0.2, size=0.2) +
  #stat_smooth(data=subset(wstidier2,
  #                        (Source == "South")), 
  #            method = "lm", formula = y ~ poly(x, 2), se = T, aes(colour = 'South'), alpha = 0.2, size=0.2) +
  theme_bw() + 
  
  scale_x_datetime(breaks = date_breaks("1 month"), labels = date_format("%b %d")) +
  #scale_x_datetime(breaks = date_breaks("2 weeks"), labels = date_format("%b %d")) +
  theme(text = element_text(size=15),
        legend.position="top"
        # axis.title.x = element_blank(),
        # axis.text.x=element_text(angle = 45, hjust = 1)
        ) +
  # geom_smooth(data=subset(ws, Source != "Outlet"), method = "lm", formula = y ~ poly(x, 2)) +
  xlab("Date") + 
  #ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  scale_y_continuous(
    expression(paste({Delta~delta}^"13","C", ' (\u2030)')),
    sec.axis = sec_axis(trans = ~ (1-((1000 + d13Co + .)/(1000+d13Co))^(1000/epsilon_lab))*100 , 
                        name = "Degradation (%)", 
                        #name = element_blank(), 
                        breaks=c(20, 40, 60, 70, 80, 85, 90, 95) )# breaks=seq(20, 120, 15))
  )  + 
  scale_color_manual(name= "Source", 
                     # Actual order:
                     # Bulk, North, Outlet, South, Valley
                      values = c("#B79F00", "#F8766D", "#00BFC4", "#C77CFF", "#00BA38"
                                 # working solution:
                                 #c("black", "#F8766D", "#00BFC4", "#DE8C00", "#00BA38"
                                 #"black", "#D55E00",  "#00BFC4",  "#B79F00", "#00BA38"
                                 # Bulk, North, outlet, South, Valley
                                 #"#D55E00", "darkgreen", "dodgerblue"
                                 ),
                     breaks=c("Bulk", "North" , "Valley" ,"South",  "Outlet"),
                     labels=c("Bulk", "North" , "Valley" ,"South",  "Outlet")
                     ) +
  
  scale_size_continuous(range = c(1, 6), breaks= c(0, 50, 100, 150, 200, 300), limits = c(0, 300))+
  annotate("rect", xmin=mindate, xmax=maxdate, ymin=0, ymax=propagatedError, alpha=0.2) 
  # scale_size_continuous(range = c(1, 3)) 

#
# Reds
# gold = "#B79F00"
# red-pink  = "#F8766D" 
# "firebrick1", 
# 'yellow', "orange1","red", 
# pink = "#F564E3" 

# Mono
# "gray35", "ghostwhite", 'gray99'

# Greens
# 'darkgreen','darkolivegreen3','darkseagreen3','darkseagreen1'
# dark green = "chartreuse4"  
# darkish freen = "#00BA38"

# Blues
# purple = "blueviolet"
# "dodgerblue", "#00BFC4" (light blue), "#619CFF" (sharp blue),
#  "deepskyblue" 

wsALL_lab
# ggplotly(wsALL_lab)
```

## Field Enrichment Plot

```{r}

mindate = as.POSIXct("2016-03-28 00:04:00" , tz = "EST") # min(wstidier2$Date)
maxdate = as.POSIXct("2016-06-27 00:01:00", tz = "EST") 
  
wsALL_field <- ggplot(data = wstidier2, aes(x = Date, y = measure, group = Source) )+ 
  # Dissolved (Outlet) trend
  stat_smooth(data=subset(wstidier2,
                          (Source == "Outlet" 
                           # & Event > 1 
                           & Type == "Dissolved (Outlet)")), 
              method = "lm", formula = y ~ poly(x, 2), se = F, aes(colour = 'Outlet'), alpha = 0.9, size=0.2) +
  # Bulk trend
  stat_smooth(data=subset(wstidier2,
                          (Source == "Bulk" )), #| Source == "South" )), 
              method = "lm", formula = y ~ poly(x, 2), se = F, aes(colour = 'Bulk'), alpha = 0.9, size=0.2) +
  # Error bars
  geom_errorbar(data=subset(wstidier2, Source == 'Bulk'), limits_DdC, size=0.2) +
  geom_errorbar(data=subset(wstidier2, Source == 'South' 
                            | Source == 'North'
                            | Source == 'Valley'
                            ), limits_DdC, size=0.1) +
  geom_errorbar(data=subset(wstidier2, Source == 'Outlet'), limits_DdC) +
  # Data points
  geom_point(data=subset(wstidier2, Source == 'Bulk'), aes(shape = Type, colour = Source), size=3) +
  geom_point(data=subset(wstidier2, (Source == 'South'
                                     | Source == 'North'
                                     | Source == 'Valley'
                                     )), aes(shape = Type, colour = Source), size=2) +
  geom_point(data=subset(wstidier2, Source == 'Outlet'), aes(shape = Type, colour = Source, size = Qmean)) +
  theme_bw() + 
  # Applications
  annotate("text", x = as.POSIXct('2016-03-28 08:04:00'), y = 0, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
  annotate("text", x = as.POSIXct('2016-04-05 00:04:00'), y = 0, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
  annotate("text", x = as.POSIXct('2016-04-13 08:04:00'), y = 0, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
  annotate("text", x = as.POSIXct('2016-05-25 08:04:00'), y = 0, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
  # Title applics
  annotate("text", x = as.POSIXct('2016-04-05 08:04:00'), y = 4.5, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
   annotate("text", x = as.POSIXct('2016-04-14 12:04:00'), y = 4.5, 
           label = as.character(expression(paste(" Applications"))), parse = T, size = 5.0) +
  scale_x_datetime(breaks = date_breaks("2 weeks"), labels = date_format("%b %d")) +
  #scale_x_datetime(breaks = date_breaks("1 month"), labels = date_format("%b %d")) +
  theme(text = element_text(size=15),
        legend.position="top"
        # axis.title.x = element_blank()
        # axis.text.x=element_text(angle = 45, hjust = 1)
        ) +
  xlab("Date") + 
  #ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  scale_y_continuous(
    expression(paste({Delta~delta}^"13","C", ' (\u2030)')),
    sec.axis = sec_axis(trans = ~ (1-((1000 + d13Co + .)/(1000+d13Co))^(1000/epsilon_field))*100 , 
                        name = element_blank(),
                        #name = "Degradation (%)", 
                        breaks=c(20, 40, 60, 70, 80, 85, 90, 95) )# breaks=seq(20, 120, 15))
  )  + 
  scale_color_manual(name= "Source", 
                     values = c("#B79F00", "#F8766D", "#00BFC4", "#C77CFF", "#00BA38"
                                # c("black", "#F8766D", "#00BFC4", "#DE8C00", "#00BA38" 
                                #  "black", "#D55E00",  "#00BFC4",  "#B79F00", "#00BA38"
                                 # Bulk, North, outlet, South, Valley
                                 ),
                     breaks=c("Bulk", "North" , "Valley" ,"South",  "Outlet"),
                     labels=c("Bulk", "North" , "Valley" ,"South",  "Outlet")
                    ) +
  scale_size_continuous(range = c(1, 6), breaks= c(0, 50, 100, 150, 200, 300), limits = c(0, 300)) +
  # scale_size_continuous(range = c(1, 3)) + 
  guides(col = guide_legend(order = 1,
                            #title=expression("Source"),
                            #title.vjust = -1,
                            nrow = 2, 
                            title.position = "top",
                            keyheight = 1.5
                            ), 
         shape=guide_legend(title=("Type"), 
                            order = 2,
                            nrow=2, 
                            title.position = "top", 
                            keyheight = 1.5, title.vjust = NULL, label.vjust = NULL
                            ), 
         size = guide_legend(order = 3, 
                             #title=expression("Mean Discharge"), 
                             title=expression("Mean Discharge (" ~m^3 / h~")" ), 
                             nrow=2, 
                             title.position = "top"
                             # title.vjust = .26
                             #keyheight = 0,
                             #label.vjust = 0
                              )) +
  annotate("rect", xmin=mindate, xmax=maxdate, ymin=0, ymax=propagatedError, alpha=0.2) 

#ggplotly(wsALL_field)
wsALL_field


```



## Join all figures
```{r}
#wsALL_lab
#wsALL_field
#wsPlot
# ggsave(wsALL, filename = "WaterSoilvsTime.png", width = 8, height = 5, units = "in", scale = 1)
# ggsave(wsALL, filename = "WaterBulkvsTime.png", width = 8, height = 5, units = "in", scale = 1)

wsALL_field_noLeg <- wsALL_field + theme(legend.position='none')
wsALL_lab_noLeg <- wsALL_lab + theme(legend.position='none')
wsAll_field_Leg  <- get_legend(wsALL_field)

labely1 = expression(epsilon ["field"])
labely2 = expression(epsilon ["lab"])

label <- substitute(paste(epsilon, " = ", epsilon_f, ", Field", epsilon, " = " , epsilon_l),
                    list(epsilon_f = signif(epsilon_field, 2), epsilon_l = signif(epsilon_lab, 2) ))

label2 <- substitute(paste(epsilon ["field"] , " = ", epsilon_f, " \u00B1 ", "0.53" ,"\u2030"),
                     list(epsilon_f = signif(epsilon_field, 3)))

label3 <- substitute(paste(epsilon ["lab"] , " = ", epsilon_l, " \u00B1 ", "0.47" ,"\u2030"),
                     list(epsilon_l = signif(epsilon_lab, 3)))
# adding label via ggdraw, in the ggdraw coordinates


wsALL <- ggdraw() +
  draw_plot(wsALL_lab_noLeg, x=0, y = 0.16, width = 1, height = 0.82) + # bottom
  draw_plot(wsALL_field_noLeg, x=0, y=.16, width = 0.928, height = .82) + # top
  draw_label(label2, x= .886, y = .10, size = 13) + # Epsilon field (bottom)
  draw_label(label3, x= .89, y = .05, size = 13) + # Epsilon lab (bottom)
  draw_label(labely1 , x= .88, y = .98, size = 14) + # Epsilon field (top)
  draw_label(labely2 , x= .95, y = .98, size = 14) + # Epsilon lab (top)
  draw_plot(wsAll_field_Leg, x=0.2, y=0.0, width = 0.50, height = 0.15)
 
wsALL 

# ggsave(wsALL, filename = "images/WaterSoilvsTime.png" , width = 10, height = 5) # , units = "in", scale = 1)


#install.packages("extrafont")
#library(extrafont)
```

# Only Bulk and Outlet Data

## Lab Enrichment Plot

```{r}
mindate = min(wstidier2$Date)
maxdate = max(wstidier2$Date)

pd <- position_dodge(width = 0.5)
limits_DdC <- aes(ymin=measure-SD, ymax=measure+SD, colour = Source)

wsALL_lab <- ggplot(data = wstidier2, aes(x = Date, y = measure, group = Source) )+ 
  geom_errorbar(data=subset(wstidier2, Source == 'Bulk'), limits_DdC, size=0.2) +
  geom_errorbar(data=subset(wstidier2, Source == 'Outlet'), limits_DdC) +
  geom_point(data=subset(wstidier2, Source == 'Outlet'), aes(shape = Type, colour = Source, size = Qmean)) +
  geom_point(data=subset(wstidier2, Source == 'Bulk'), aes(shape = Type, colour = Source)) +
  
  # Water
  stat_smooth(data=subset(wstidier2,
                          (Source == "Outlet" 
                           # & Event > 1 
                           & Type == "Dissolved (Outlet)")), 
              method = "lm", formula = y ~ poly(x, 2), se = F, aes(colour = 'Outlet'), alpha = 0.2, size=0.2) +
  # North
  stat_smooth(data=subset(wstidier2,
                          (Source == "Bulk" )), #| Source == "South" )), 
              method = "lm", formula = y ~ poly(x, 2), se = F, aes(colour = 'Bulk'), alpha = 0.2, size=0.2) +
  #stat_smooth(data=subset(wstidier2,
  #                        (Source == "South")), 
  #            method = "lm", formula = y ~ poly(x, 2), se = T, aes(colour = 'South'), alpha = 0.2, size=0.2) +
  theme_bw() + 
  
  scale_x_datetime(breaks = date_breaks("1 month"), labels = date_format("%b %d")) +
  #scale_x_datetime(breaks = date_breaks("2 weeks"), labels = date_format("%b %d")) +
  theme(text = element_text(size=15),
        legend.position="top"
        # axis.title.x = element_blank(),
        # axis.text.x=element_text(angle = 45, hjust = 1)
        ) +
  # geom_smooth(data=subset(ws, Source != "Outlet"), method = "lm", formula = y ~ poly(x, 2)) +
  xlab("Date") + 
  #ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  scale_y_continuous(
    expression(paste({Delta~delta}^"13","C", ' (\u2030)')),
    sec.axis = sec_axis(trans = ~ (1-((1000 + d13Co + .)/(1000+d13Co))^(1000/epsilon_lab))*100 , 
                        name = "Degradation (%)", 
                        #name = element_blank(), 
                        breaks=c(20, 40, 60, 70, 80, 85, 90, 95) )# breaks=seq(20, 120, 15))
  )  + 
  scale_color_manual(name= "Source", 
                     # Actual order:
                     # Bulk, North, Outlet, South, Valley
                      values = c("#B79F00", "#00BFC4"
                                 # working solution:
                                 #c("black", "#F8766D", "#00BFC4", "#DE8C00", "#00BA38"
                                 #"black", "#D55E00",  "#00BFC4",  "#B79F00", "#00BA38"
                                 # Bulk, North, outlet, South, Valley
                                 #"#D55E00", "darkgreen", "dodgerblue"
                                 ),
                     breaks=c("Bulk", "Outlet"),
                     labels=c("Bulk", "Outlet")
                     ) +
  
  scale_size_continuous(range = c(1, 6), breaks= c(0, 50, 100, 150, 200, 300), limits = c(0, 300))+
  annotate("rect", xmin=mindate, xmax=maxdate, ymin=0, ymax=propagatedError, alpha=0.2) 
  # scale_size_continuous(range = c(1, 3)) 

#
# Reds
# gold = "#B79F00"
# red-pink  = "#F8766D" 
# "firebrick1", 
# 'yellow', "orange1","red", 
# pink = "#F564E3" 

# Mono
# "gray35", "ghostwhite", 'gray99'

# Greens
# 'darkgreen','darkolivegreen3','darkseagreen3','darkseagreen1'
# dark green = "chartreuse4"  
# darkish freen = "#00BA38"

# Blues
# purple = "blueviolet"
# "dodgerblue", "#00BFC4" (light blue), "#619CFF" (sharp blue),
#  "deepskyblue" 

wsALL_lab
# ggplotly(wsALL_lab)
```

## Field Enrichment Plot

```{r}

mindate = as.POSIXct("2016-03-28 00:04:00" , tz = "EST") # min(wstidier2$Date)
maxdate = as.POSIXct("2016-06-27 00:01:00", tz = "EST") 
  
wsALL_field <- ggplot(data = wstidier2, aes(x = Date, y = measure, group = Source) )+ 
  # Dissolved (Outlet) trend
  stat_smooth(data=subset(wstidier2,
                          (Source == "Outlet" 
                           # & Event > 1 
                           & Type == "Dissolved (Outlet)")), 
              method = "lm", formula = y ~ poly(x, 2), se = F, aes(colour = 'Outlet'), alpha = 0.9, size=0.2) +
  # Bulk trend
  stat_smooth(data=subset(wstidier2,
                          (Source == "Bulk" )), #| Source == "South" )), 
              method = "lm", formula = y ~ poly(x, 2), se = F, aes(colour = 'Bulk'), alpha = 0.9, size=0.2) +
  # Error bars
  geom_errorbar(data=subset(wstidier2, Source == 'Bulk'), limits_DdC, size=0.2) +
  geom_errorbar(data=subset(wstidier2, Source == 'Outlet'), limits_DdC) +
  # Data points
  geom_point(data=subset(wstidier2, Source == 'Bulk'), aes(shape = Type, colour = Source), size=3) +
  geom_point(data=subset(wstidier2, Source == 'Outlet'), aes(shape = Type, colour = Source, size = Qmean)) +
  theme_bw() + 
  # Applications
  annotate("text", x = as.POSIXct('2016-03-28 08:04:00'), y = 0, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
  annotate("text", x = as.POSIXct('2016-04-05 00:04:00'), y = 0, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
  annotate("text", x = as.POSIXct('2016-04-13 08:04:00'), y = 0, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
  annotate("text", x = as.POSIXct('2016-05-25 08:04:00'), y = 0, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
  # Title applics
  annotate("text", x = as.POSIXct('2016-04-05 08:04:00'), y = 3.5, 
           label = as.character(expression(paste( "+"))), parse = T, size = 6.0) +
   annotate("text", x = as.POSIXct('2016-04-14 12:04:00'), y = 3.5, 
           label = as.character(expression(paste(" Applications"))), parse = T, size = 5.0) +
  scale_x_datetime(breaks = date_breaks("2 weeks"), labels = date_format("%b %d")) +
  #scale_x_datetime(breaks = date_breaks("1 month"), labels = date_format("%b %d")) +
  theme(text = element_text(size=15),
        legend.position="top"
        # axis.title.x = element_blank()
        # axis.text.x=element_text(angle = 45, hjust = 1)
        ) +
  xlab("Date") + 
  #ylab(expression(paste({Delta~delta}^"13","C", ' (\u2030)'))) +
  scale_y_continuous(
    expression(paste({Delta~delta}^"13","C", ' (\u2030)')),
    sec.axis = sec_axis(trans = ~ (1-((1000 + d13Co + .)/(1000+d13Co))^(1000/epsilon_field))*100 , 
                        name = element_blank(),
                        #name = "Degradation (%)", 
                        breaks=c(20, 40, 60, 70, 80, 85, 90, 95) )# breaks=seq(20, 120, 15))
  )  + 
  scale_color_manual(name= "Source", 
                     # Bulk, North, Outlet, South, Valley
                     values = c("#B79F00", "#00BFC4"
                                # c("black", "#F8766D", "#00BFC4", "#DE8C00", "#00BA38" 
                                #  "black", "#D55E00",  "#00BFC4",  "#B79F00", "#00BA38"
                                 # Bulk, North, outlet, South, Valley
                                 ),
                     breaks=c("Bulk", "Outlet"),
                     labels=c("Bulk", "Outlet")
                    ) +
  scale_size_continuous(range = c(1, 6), breaks= c(0, 50, 100, 150, 200, 300), limits = c(0, 300)) +
  # scale_size_continuous(range = c(1, 3)) + 
  guides(col = guide_legend(order = 1,
                            #title=expression("Source"),
                            #title.vjust = -1,
                            nrow = 2, 
                            title.position = "top",
                            keyheight = 1.5
                            ), 
         shape=guide_legend(title=("Type"), 
                            order = 2,
                            nrow=2, 
                            title.position = "top", 
                            keyheight = 1.5, title.vjust = NULL, label.vjust = NULL
                            ), 
         size = guide_legend(order = 3, 
                             #title=expression("Mean Discharge"), 
                             title=expression("Mean Discharge (" ~m^3 / h~")" ), 
                             nrow=2, 
                             title.position = "top"
                             # title.vjust = .26
                             #keyheight = 0,
                             #label.vjust = 0
                              )) +
  annotate("rect", xmin=mindate, xmax=maxdate, ymin=0, ymax=propagatedError, alpha=0.2) 

#ggplotly(wsALL_field)
wsALL_field


```



## Join all figures
```{r}
#wsALL_lab
#wsALL_field
#wsPlot
# ggsave(wsALL, filename = "WaterSoilvsTime.png", width = 8, height = 5, units = "in", scale = 1)
# ggsave(wsALL, filename = "WaterBulkvsTime.png", width = 8, height = 5, units = "in", scale = 1)

wsALL_field_noLeg <- wsALL_field + theme(legend.position='none')
wsALL_lab_noLeg <- wsALL_lab + theme(legend.position='none')
wsAll_field_Leg  <- get_legend(wsALL_field)

labely1 = expression(epsilon ["field"])
labely2 = expression(epsilon ["lab"])

label <- substitute(paste(epsilon, " = ", epsilon_f, ", Field", epsilon, " = " , epsilon_l),
                    list(epsilon_f = signif(epsilon_field, 2), epsilon_l = signif(epsilon_lab, 2) ))

label2 <- substitute(paste(epsilon ["field"] , " = ", epsilon_f, " \u00B1 ", "0.53" ,"\u2030"),
                     list(epsilon_f = signif(epsilon_field, 3)))

label3 <- substitute(paste(epsilon ["lab"] , " = ", epsilon_l, " \u00B1 ", "0.47" ,"\u2030"),
                     list(epsilon_l = signif(epsilon_lab, 3)))
# adding label via ggdraw, in the ggdraw coordinates


wsALL <- ggdraw() +
  draw_plot(wsALL_lab_noLeg, x=0, y = 0.16, width = 1, height = 0.81) + # bottom
  draw_plot(wsALL_field_noLeg, x=0, y=.16, width = 0.928, height = .81) + # top
  draw_label(label2, x= .886, y = .10, size = 13) + # Epsilon field (bottom)
  draw_label(label3, x= .89, y = .05, size = 13) + # Epsilon lab (bottom)
  draw_label(labely1 , x= .88, y = .98, size = 14) + # Epsilon field (top)
  draw_label(labely2 , x= .95, y = .98, size = 14) + # Epsilon lab (top)
  draw_plot(wsAll_field_Leg, x=0.2, y=0.0, width = 0.50, height = 0.15)
 
wsALL 

# ggsave(wsALL, filename = "images/WaterSoilvsTime_BulkOnly.png" , width = 10, height = 5) # , units = "in", scale = 1)


#install.packages("extrafont")
#library(extrafont)
```
